% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{abundance}
\alias{abundance}
\title{abundance}
\usage{
abundance(x, taxLev = "Family", unassigned = FALSE)
}
\arguments{
\item{x}{result of the function aggregatoR}

\item{taxLev}{taxonomic level on which the calculation has to be made.}

\item{unassigned}{Does unassigned taxa need to be taken into account? If yes set unassigned to TRUE, otherwise FALSE.}
}
\description{
This function allows the calculation of abundances with the possibility to exclude unassigned taxa at the desired taxonomic level.
}
\details{
If unassigned is set to TRUE abundances are equal among taxonomic levels.
}
\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
abundance(data.agR, taxLev = "Family" )
abundance(data.agR, taxLev = "Family" , unassigned = TRUE)
}
\seealso{
\code{\link{aggregatoR}}
}
\keyword{abundance}
