% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_classes.R
\name{abundance_classes}
\alias{abundance_classes}
\title{transform abundance-coverage data to classes}
\usage{
abundance_classes(
  x,
  abundance_classes = c(0.1, 1, 10, 50),
  class_labels = NULL
)
}
\arguments{
\item{x}{A \code{matrix} or a \code{data.frame}.}

\item{abundance_classes}{A numeric vector containing the values for separating classes.}

\item{class_labels}{A numeric vector containing the labels of the classes. Default to the number of abundance classes plus 1.}
}
\description{
From abundance or coverage data it returns classes.
}
\details{
Some indices require abundance or coverage to be transformed in abundance or coverage classes.
}
\examples{

data(mi_prin)
abundance_classes(mi_prin)
}
