% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abuTax.R
\name{abuTax}
\alias{abuTax}
\title{abuTax - deprecated}
\usage{
abuTax(x, taxa = NULL, rel = FALSE)
}
\arguments{
\item{x}{result of the function aggregatoR.}

\item{taxa}{a taxon or a vector of taxa.}

\item{rel}{if TRUE calculates relative abundance. Default to FALSE.}
}
\description{
\Sexpr[results=rd, stage=render]{ lifecycle::badge("deprecated") }

This function calculates the absolute or relative abundance of a taxon or of a set taxa.
}
\details{
This function does not check for parent-child pairs. For instance if the vector of taxa contains both the genus \emph{Baetis} and the
family Baetidae, the abundance of the genus \emph{Baetis} will be double counted. Check the function \code{\link{ambiguousSolver}} for a solution
to this problem.
}
\seealso{
\code{\link{get_taxa_abundance}}
}
\keyword{aggregatoR}
