% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBiasToTraits.R
\name{addBiasToTraits}
\alias{addBiasToTraits}
\title{Add bias to duplicated fuzzy traits}
\usage{
addBiasToTraits(x, colB = NULL, SD = 0.001, traceB = FALSE)
}
\arguments{
\item{x}{results of \code{traitsMean} or a \code{data.frame} with a similar structure.}

\item{colB}{A vector that contains the number of modalities for each trait}

\item{SD}{the amount of bias.}

\item{traceB}{if \code{TRUE} returns a vector of taxa with duplicated traits.}
}
\description{
Add bias to duplicated fuzzy traits
}
\details{
This function works by adding a small bias to duplicated fuzzy traits.
The bias is added with the \code{rnorm} function.
}
