% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_taxa.R
\name{aggregate_taxa}
\alias{aggregate_taxa}
\title{aggregate_taxa}
\usage{
aggregate_taxa(x, FUN = sum)
}
\arguments{
\item{x}{Result of \code{as_biomonitor()}.}

\item{FUN}{The function to be applied for aggregating at higher taxonomic levels.
Must be sum for both abundances and presence-absence data.
Default to \code{sum()}.}
}
\description{
This function aggregates taxa at various taxonomic levels. Preparing data with \code{aggregate_taxa()} is necessary for
further calculations.
}
\examples{
data(macro_ex)
data_bio <- as_biomonitor(macro_ex)
data_agr <- aggregate_taxa(data_bio)

# example for macrophytes
data(oglio)

oglio_asb <- as_biomonitor(oglio, group = "mf")
oglio_agg <- aggregate_taxa(oglio_asb)
richness(oglio_agg, tax_lev = "Species")
richness(oglio_agg, tax_lev = "Genus")
richness(oglio_agg, tax_lev = "Family")
}
\seealso{
\link{as_biomonitor}
}
\keyword{aggregate_taxa}
