% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregatoR.R
\name{aggregatoR}
\alias{aggregatoR}
\title{aggregatoR - deprecated}
\usage{
aggregatoR(x, FUN = sum)
}
\arguments{
\item{x}{results of function asBiomonitor}

\item{FUN}{the function to be applied for aggregating to higher taxonomic levels.
Must be sum for both abundances and presence-absence data.
Default to \code{sum}. Change if you know what you are doing.}
}
\description{
\Sexpr[results=rd, stage=render]{ lifecycle::badge("deprecated") }

This function prepares data for further calculations.
}
\seealso{
\code{\link{aggregate_taxa}}
}
\keyword{aggregatoR}
