% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allrich.R, R/richness.R
\name{allrich}
\alias{allrich}
\alias{richness}
\title{Function for calculating richness at different taxonomic levels}
\usage{
allrich(x)

richness(x, tax_lev = "Taxa")
}
\arguments{
\item{x}{result of the function aggregatoR}

\item{tax_lev}{taxonomic level on which the calculation has to be made.}
}
\description{
This function allows the calculation of richness at different taxonomic levels and of the total abundance.
}
\details{
The function \code{richness} returns a warning meassage if unassigned taxa are detected at the desired taxonomic level. Unassigned taxa can be present
because some taxa are entered at higher taxonomic levels than those at which the calculation is performed.
For instance this problem appears when the calculation is done at genus level and some taxa has been entered at family level.
Unassigned taxa can be present also because of missing levels in the reference database at the desired taxonomic level.
This problem is present, for example, for Mollusca in the default database for macroinvertebrates.
Mollusca orders are not present because there is no consensus among taxonomist (see the \href{https://www.molluscabase.org/index.php}{www.molluscabase.org} website).
}
\section{Functions}{
\itemize{
\item \code{richness}: Taxonomic richness
}}

\examples{
data(macro_ex)
data_bio <- as_biomonitor(macro_ex)
data_agr <- aggregate_taxa(data_bio)
richness(data_agr, tax_lev = "Family")
}
\seealso{
\code{\link{aggregatoR}}
}
\keyword{allrich,}
\keyword{richness}
