% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiguousSolver.R
\name{ambiguousSolver}
\alias{ambiguousSolver}
\title{ambiguousSolver}
\usage{
ambiguousSolver(x, method = "MCWP")
}
\arguments{
\item{x}{result of the function aggregatoR.}

\item{method}{only methods \code{RPKC} and \code{MCWP} are currently implemented in \code{biomonitoR}. See details.}
}
\description{
\Sexpr[results=rd, stage=render]{ lifecycle::badge("deprecated") }

Solve the ambiguous assignment of taxa.
}
\details{
Taxonomic dataset often contains ambiguous taxa due to the difficulites in identify damaged and juvenile organisms.
For instance Chironominae and Chironomidae could be reported in the same sample. Several techinques has been
proposed to solve this issue (see Cuffney et al., 2007).
Currently ambiguousSolver two only one algorithms to solve the problem of ambiguous assignments.
\code{RPKC} means Remove Parents - Keep Children. It works by removing all the higher taxonomic levels while keeping children.
\code{MCWP} means Merge Children with Parents. It works by assignign the abundances of lower taxonomic levels to those of higher taxonomic levels.
}
\references{
Cuffney, T. F., Bilger, M. D., & Haigler, A. M. (2007). Ambiguous taxa: effects on the characterization and interpretation of invertebrate assemblages. Journal of the North American Benthological Society, 26(2), 286-307.
}
\seealso{
\code{\link{aggregatoR}}
}
\keyword{aggregate_taxa}
