% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asBiomonitor.R
\name{asBiomonitor}
\alias{asBiomonitor}
\title{asBiomonitor}
\usage{
asBiomonitor(x, dfref = NULL, overwrite = F)
}
\arguments{
\item{x}{a data.frame as specified in details}
}
\description{
This function prepares data for further calculations.
}
\details{
data.frame must have a column called "Taxa" where put species, genus or family names. See data(macro_ex) for an example dataset.\cr
asBiomonitor check the correctness of taxa names in the data.frame provided by the user. If names are correct the function will process the data.frame to a biomonitor object, else it suggest to rename the wrong names with the function \code{\link{rename}}.
}
\examples{
data(macro_ex)
asBiomonitor(macro_ex)
}
\seealso{
\code{\link{rename}}
}
\keyword{asBiomonitor}
