% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioco.R
\name{bioco}
\alias{bioco}
\title{bioco}
\usage{
bioco(x, alien = NULL, dfref = NULL, digits = 2)
}
\arguments{
\item{x}{Result of function \code{aggregate_taxa()}.}

\item{alien}{a vector containing the alien taxa name. Only taxa up to family level will be considered.}

\item{dfref}{reference database.
If a default reference database was used use \code{mi} for macroinvertebrates or \code{mf} for macrophytes.
Use the same reference database you used for \code{as_biomonitor()} otherwise.}

\item{digits}{number of decimal places, default to 2 as in the original work of Arbaciauskas et al. (2008).}
}
\description{
bioco function provides site-specific biocontamination index (SBCI), abundance contamination index (ACI) and richness contamination index (RCI) at family rank according to those proposed by Arbaciauskas et al. (2008).
}
\details{
Biocontamination of sampling sites was assessed using a site-specific biocontamination index (SBCI) derived from two metrics: abundance contamination index (ACI) and richness contamination index (RCI) at family/ordinal rank. These indices were calculated as:
\deqn{ACI = Na/Nt}
where Na and Nt are numbers of specimens of alien taxa and total specimens in a sample, respectively, and
\deqn{RCI = Ta/Tt}
where Ta is the total number of alien families/orders, and Tt is the total number of identified families/orders. With values of ACI and RCI, the site-specific biocontamination index (SBCI) can then be derived from an easy double entry matrix (Arbaciauskas et al. 2008). Five classes of biocontamination ranging from 0 ("no" contamination) to 4 ("severe" contamination) are defined.
A global alien species dataset cannot be provided because alien species may vary among different biogeographical regions and countries. For this reason users need to define and properly upload their own alien reference database based on the studied area they want to study.
Applications and examples of these indices are available in Cuk et al. (2019) for Croatian rivers and MacNeil et al. 2010 (The Isle of Man).
}
\examples{
data(macro_ex)
data_bio <- as_biomonitor(macro_ex)
data_agr <- aggregate_taxa(data_bio)
# Toy example:
alien <- c("Laccobius", "Setodes bulgaricus", "Caenidae")
bioco(data_agr, alien = alien, dfref = "mi")
}
\references{
Arbaciauskas, K., Semenchenko, V., Grabowski, M., Leuven, R.S.E.W., Paunovic, M., Son, M.O., Csanyi, B., Gumuliauskaite, S., Konopacka, A., Nehring, S., Van der Velde, G., Vezhnovetz, V., Panov, V.E. (2008). Assessment of biocontamination of benthic macroinvertebrate communities in European inland waterways. Aquatic Invasions 3 (2): 211-230.

Cuk, R., Milisa, M., Atanackovic, A., Dekic, S., Blazekovic, L., & Zganec, K. (2019). Biocontamination of benthic macroinvertebrate assemblages in Croatian major rivers and effects on ecological quality assessment. Knowledge & Management of Aquatic Ecosystems, (420), 11.

MacNeil, C., Briffa, M., Leuven, R.S., Gell, F.R. and Selman, R. (2010). An appraisal of a biocontamination assessment method for freshwater macroinvertebrate assemblages; a practical way to measure a significant biological pressure? Hydrobiologia 638:151-159
}
\seealso{
\link{aggregate_taxa}
}
\keyword{aggregate_taxa}
