% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bqies.R
\name{bqies}
\alias{bqies}
\title{bqies}
\usage{
bqies(d, method = "bqies")
}
\arguments{
\item{d}{results of function aggregatoR}
}
\description{
This function calculates the Benthic Quality Index for Italian Lakes. Boggero et al. (2016), Italian classification method for macroinvertebrates in lakes. Method summary.
}
\details{
bqies is calculated....
}
\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
bqies(data.agR)
bqies(data.agR, method = "i")
}
\references{
Boggero A., Zaupa S., Cancellario T., Lencioni V., Marziali L., Rossaro B., (2016), Italian classification method for macroinvertebrates in lakes. Method summary. REPORT
}
\seealso{
\code{\link{aggregatoR}}
}
\keyword{aggregatoR}
