% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_taxa.R
\name{combine_taxa}
\alias{combine_taxa}
\title{combine_taxa}
\usage{
combine_taxa(x, ntaxa = 2, tax_lev = "Taxa")
}
\arguments{
\item{x}{Result of the function \code{aggregate_taxa()}.}

\item{ntaxa}{Rumber of taxa.}

\item{tax_lev}{Taxonomic level on which the calculation has to be made.}
}
\description{
This function returns all the combinations of n taxa at the desired taxonomic resolution.
}
\details{
This function is intended to help the user to identify the best subset of taxa that correlates with environmental variables.
Metrics based on a specific group of taxa (EPT, 1-GOLD, etc) are currently used in biomonitoring. They rely on available information on the responses
of the selected taxa to target strssors. The relationship between an environmental variable and a subset of taxa could exist and not detected due to the scarce
availaibility of autoecological data. With more than 4 taxa the calculations should become infeasible, expecially when the number
of taxa in the user dataset is high.
}
\examples{
data(macro_ex)
data_bio <- as_biomonitor(macro_ex)
data_agr <- aggregate_taxa(data.bio)
combine_taxa(data_agr, tax_lev = "Family")
}
\seealso{
\code{\link{aggregate_taxa}}
}
\keyword{aggregatoR}
