% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combTaxa.R
\name{combTaxa}
\alias{combTaxa}
\title{combTaxa}
\usage{
combTaxa(x, ntaxa = 2, taxLev = "Family", rel = F)
}
\arguments{
\item{x}{results of function aggregatoR.}

\item{rel}{if TRUE calculates relative abundance. default =F.}

\item{taxa}{a Taxon or a vector of taxa.}
}
\description{
This function calculates the absolute or relative abundance of a Taxon or of a set Taxa.
}
\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
combTaxa(data.agr, taxLev = "Family")
}
\seealso{
\code{\link{aggregatoR}}
}
\keyword{aggregatoR}
