% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertobiotic.R, R/convertovegan.R
\name{convertobiotic}
\alias{convertobiotic}
\alias{convertovegan}
\title{convertovegan}
\usage{
convertobiotic(x)

convertovegan(x, taxLev = "Family")
}
\arguments{
\item{x}{results of aggregatoR function}

\item{taxLev}{taxonomic level of interest. Possible choices are Phylum, Class, Subclass, Order, Family, Subfamily, Tribus, Genus, Species, Subspecies, Taxa}
}
\description{
An utility function to export data to other formats.
}
\details{
\code{convertovegan} converts data to a vegan style format, with sites on rows and taxa on columns. \code{convertovegan} converts data to
the format of biotic package. For Oligochaeta, if both family and above family level taxonomic information are concurrently present,
information will be provided at Oligochaeta level.
}
\section{Functions}{
\itemize{
\item \code{convertobiotic}: convert data to biotic format
}}

\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
data.cv <- convertovegan(data.agR, taxLev = "Family")
}
\seealso{
\code{\link{asBiomonitor}}
}
\keyword{convertovegan}
