% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{defunct}
\alias{defunct}
\alias{asBiomonitor}
\alias{aggregatoR}
\alias{abuTax}
\alias{addBiasToTraits}
\alias{ambiguousSolver}
\alias{combTaxa}
\alias{convertovegan}
\alias{convertobiotic}
\alias{totInfo}
\alias{manageTraits}
\alias{quickRename}
\alias{refFromTree}
\alias{sampleTraits}
\alias{selectPcoaAxes}
\alias{showscores}
\alias{ricTax}
\alias{traitsMean}
\alias{traitScaling}
\title{Defunct functions}
\usage{
asBiomonitor(.variables, drop = FALSE)

aggregatoR(.variables, drop = FALSE)

abuTax(.variables, drop = FALSE)

addBiasToTraits(.variables, drop = FALSE)

ambiguousSolver(.variables, drop = FALSE)

combTaxa(.variables, drop = FALSE)

convertovegan(.variables, drop = FALSE)

convertobiotic(.variables, drop = FALSE)

totInfo(.variables, drop = FALSE)

manageTraits(.variables, drop = FALSE)

quickRename(.variables, drop = FALSE)

refFromTree(.variables, drop = FALSE)

sampleTraits(.variables, drop = FALSE)

selectPcoaAxes(.variables, drop = FALSE)

showscores(.variables, drop = FALSE)

ricTax(.variables, drop = FALSE)

traitsMean(.variables, drop = FALSE)

traitScaling(.variables, drop = FALSE)
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("defunct")}
Executing these functions will tell you which function replaces them.
}
\keyword{internal}
