% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dehli.R
\name{dehli}
\alias{dehli}
\title{Drought Effect of Habitat Loss on Invertebrates}
\usage{
dehli(x, method = "chadd_2017", agg = TRUE, exceptions = NULL, traceB = FALSE)
}
\arguments{
\item{x}{Result of \code{aggregate_taxa()}.}

\item{method}{The implementation of BMWP needed to calculate ASPT. Possible choices are \code{a} (Armitage et al. 1983), \code{uk} (Davy-Bowker et al. 2010), \code{spa} (MAGRAMA 2011), \code{ita} (Buffagni et al . 2014).
Users can provide their own data.frame (see examples) with a column called \emph{Taxon} and the column of scores called \emph{Scores}.}

\item{agg}{This option allows the composite family approach. It can be \code{FALSE}, \code{TRUE} or a \code{data.frame}.
If \code{FALSE} no aggregation will be performed, while if \code{TRUE} aggregation will be performed according to the rules described in \emph{Details}.
A \code{data.frame} containing the aggregation rules can be provided by the user.
This \code{data.frame} needs a column called \emph{Taxon} containing the taxon to aggregate and a column called \emph{Correct_Taxon} with the aggregation specifications.
\code{agg} cannot be \code{TRUE} when a \code{data.frame} is provided as method.}

\item{exceptions}{Taxa that need to be excluded from the calculation.
This option can be useful, for instance, to exclude an alien species belonging to an autochthonous family.}

\item{traceB}{If set to \code{TRUE} a list as specified below will be returned.}
}
\value{
If \code{traceB} is set to \code{TRUE} a list with the following elements will be returned:
\itemize{
\item \code{results} Result of \code{dehli()}.
\item \code{taxa_df} The \code{data.frame} used for the calculation containing the abundance of taxa receiving a score.
\item \code{composite_taxa} Taxa aggregated following the aggregation rules when \code{agg} is not \code{NULL}.
\item \code{exceptions} A \code{data.frame} containing the changes made by excluding the taxa included in \code{exceptions}.
\item \code{parent_child_pairs} For instance in Spanish ASPT both \emph{Ferrissia} and Planorbidae receive a score.
Abundances of the higher taxonomic level need therefore to be adjusted by subtracting the abundances of the lower taxonomic level.
}
}
\description{
This function calculates the \emph{Average Score Per Taxon} index following Armitage et al. (1983), Davy-Bowker et al. (2007) and Alba-Tercedor & Sanchez-Ortega (1988) implementations.
}
\examples{
data(macro_ex)
data_bio <- as_biomonitor(macro_ex)
data_agr <- aggregate_taxa(data_bio)
dehli(data_agr)
dehli(data_agr)
}
\references{
Chadd, R. P., England, J. A., Constable, D., Dunbar, M. J., Extence, C. A., Leeming, D. J., Murray-Bligh Wood J.A., P. J. (2017). An index to track the ecological effects of drought development and recovery on riverine invertebrate communities. Ecological Indicators, 82, 344-356.

The \code{dehli()} function automatically check for parent-child pairs in the scoring system, see the Value section for a definition.
All the information used for \code{dehli()} calculation can be retrieved with the function \code{\link{show_scores}}.
}
\seealso{
\link{aggregate_taxa} \link{bmwp}
}
\keyword{dehli}
