% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epsi.R
\name{epsi}
\alias{epsi}
\title{Empyrically-weighted Proportion of Sediment-sensitive Invertebrates index}
\usage{
epsi(
  x,
  method = "uk",
  agg = FALSE,
  abucl = c(1, 9, 99, 999),
  exceptions = NULL,
  traceB = FALSE
)
}
\arguments{
\item{x}{Results of \code{aggregate_taxa()}.}

\item{method}{The only avialble method is \code{uk}.
Users can provide their own \code{data.frame} (see examples) with a column called \emph{Taxon} and the column of scores called \emph{Scores}.}

\item{agg}{Default to \code{FALSE}. If set to \code{TRUE} \code{epsi()} will use the aggreation rules described in the details.
It can also be a \code{data.frame} provided by the user containing the specification on how to aggregate taxa. This \code{data.frame} needs a column called \emph{Taxon}
containing the taxon to aggregate and a column called \emph{Correct_Taxon} with the aggregation specifications. Used when users want to aggregate some taxonomic levels while using their own \code{data.frame} of scores (provided via \code{method}).}

\item{abucl}{Log abundance categories. Tresholds are set to 1, 9, 99, 999.}

\item{exceptions}{Taxa that need to be exluded from the calculation.
This option can be useful, for instance, to exclude an alien species belonging to an autochthonous family.
\code{agg} cannot be \code{TRUE} when a data.frame is provided as \code{method}.}

\item{traceB}{If set to \code{TRUE} a list as specified below will be returned.}
}
\value{
If \code{traceB} is set to \code{TRUE} a list with the following elements will be returned:
\itemize{
\item \code{results} Results of \code{epsi()}.
\item \code{taxa_df} The \code{data.frame} used for the calculation containing the abundance of taxa receiving a score.
\item \code{epsi_df} The \code{data.frame} used for the calculation containing scores and abundance classes for each site.
\item \code{composite_taxa} Taxa aggregated following the aggregation of the default method or set in \code{agg}.
\item \code{exceptions} A \code{data.frame} containing the containing changes made by excluding the taxa included in \code{exceptions}.
\item \code{parent_child_pairs} Parent-child pairs are not present in  the default implementation of epsi. A \code{data.frame} if both a  taxon and a parent
taxon receive a score.
}
}
\description{
\Sexpr[results=rd, stage=render]{ lifecycle::badge("maturing") }

This function calculates the Empyrically-weighted Proportion of Sediment-sensitive Invertebrates index (ePSI) according to the most recent version used in UK.
}
\details{
\code{epsi()} implementation take into account composite taxa as follow:
\enumerate{
\item Tipulidae (inc. Limoniidae, Pediciidae & Cylindrotomidae)
\item Siphlonuridae (inc. Ameletidae)
\item Hydrophilidae (inc. Georissidae, Helophoridae & Hydrochidae)
}

\code{epsi()} automatically check for parent-child pairs in the scoring system, see the return section for a definition.
All the information used for \code{epsi()} calculation can be retrieved with the function \code{\link{show_scores}}.
}
\section{Acknowledgements}{
 We thank Carol Fitzpatrick, Richard Chadd, Judy England and Rachel Stubbington for providing us with the most updated ePSI scores and algorithms.
}

\examples{
data(macro_ex)
data_bio <- as_biomonitor(macro_ex)
data_agr <- aggregate_taxa(data_bio)
epsi(data_agr)

# provide your own score sistem. Scores and aggregation rules are for example purpose only.

epsi_scores <- data.frame(
  Taxon = c("Ephemerellidae", "Leuctridae", "Chironomidae"),
  Scores = c(0.1, 0.5, 0.2)
)
epsi_acc <- data.frame(Taxon = "Ephemerellidae", Correct_Taxon = "Chironomidae")

epsi(data_agr, method = epsi_scores, agg = epsi_acc, traceB = TRUE)
}
\references{
Turley MD, Bilotta GS, Chadd RP, Extence CA, Brazier RE, Burnside NG, Pickwell AG. 2016. A sediment-specific family-level biomonitoring tool to identify the impacts of fine sediment in temperate rivers and streams. Ecological Indicators 70, 151-165.

Turley MD, Bilotta GS, Krueger T, Brazier RE, Extence CA. 2015. Developing an improved biomonitoring tool for fine sediment: combining expert knowledge and empirical data. Ecological indicators 54, 82-86.
}
\seealso{
\code{\link{aggregate_taxa}}
}
\keyword{epsi}
