% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ept.R
\name{ept}
\alias{ept}
\title{ept}
\usage{
ept(x, taxLev = "Family")
}
\arguments{
\item{x}{results of function aggregatoR}

\item{taxLev}{the taxonomic level for calculating EPT richness.}
}
\description{
This function calculates the number of Ephemeroptera, Plecotera and Trichoptera (ept) taxa.
}
\details{
The parameter taxLev can be "Species", "Genus", "Family" or "Order".
}
\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
ept(data.agR)
}
\seealso{
\code{\link{aggregatoR}}
}
\keyword{ept}
