% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_divs.R
\name{f_divs}
\alias{f_divs}
\title{Functional diversity}
\usage{
f_divs(
  x,
  trait_db = NULL,
  tax_lev = "Taxa",
  type = NULL,
  traitSel = FALSE,
  col_blocks = NULL,
  distance = "gower",
  zerodist_rm = FALSE,
  traceB = FALSE,
  correction = "none",
  set_param = NULL
)
}
\arguments{
\item{x}{Result of \code{aggregate_taxa()}.}

\item{trait_db}{A trait dataset. Can be a \code{data.frame} or a \code{dist} object.
Taxonomic level of the traits dataset must match those of the taxonomic dataset.
No automatic check is done.}

\item{tax_lev}{Character string giving the taxonomic level used to retrieve
trait information. Possible levels are \code{Taxa}, \code{Species}, \code{Genus},
\code{Family} as returned by \code{aggregate_taxa()}.}

\item{type}{The type of variables speciefied in \code{trait_db}.
Must be one of \code{F}, fuzzy, or \code{C}, continuous.
If more control is needed please consider to provide \code{trait_db} as a \code{dist} object.
It works only when \code{trait_db} is a \code{data.frame}, otherwise ingored.}

\item{traitSel}{Interactively select traits.}

\item{col_blocks}{A vector that contains the number of modalities for each trait.
Not needed when \code{euclidean} distance is used.}

\item{distance}{To be used to compute functional distances, \code{euclidean} or \code{gower}. Default to \code{gower}. See details.}

\item{zerodist_rm}{If \code{TRUE} aggregates taxa with the same traits.}

\item{traceB}{If \code{TRUE} returns a list as specified in details.}

\item{correction}{Correction methods for negative eigenvalues, can be one of \code{none}, \code{lingoes}, \code{cailliez}, \code{sqrt} and \code{quasi}.
Ignored when type is set to \code{C}.}

\item{set_param}{A list of parameters for fine tuning the calculations.
\code{max_nbdim} set the maximum number of dimension for evaluating the quality of the functional space.
\code{prec} can be \code{Qt} or \code{QJ}, please refere to the \code{convhulln} documentation for more information.
Deafault to \code{QJ}, less accurate but less prone to errors.
\code{tol} a tolerance threshold for zero, see the function \code{is.euclid}, \code{lingoes} and \code{cailliez} from the \code{ade4} for more details. Default to 1e-07.
If \code{cor.zero} is \code{TRUE}, zero distances are not modified. see the function \code{is.euclid}, \code{lingoes} and \code{cailliez} from the \code{ade4} for more details. Default to \code{TRUE}.}
}
\value{
a vector with fuzzy functional richness results.
\enumerate{
\item \code{results} Results of \code{f_divs()}.
\item \code{traits} A \code{data.frame} containing the traits used for the calculations.
\item \code{taxa} A \code{data.frame} containing the taxa used for the calculations.
\item \code{correction} The type of correction used.
\item \code{NA_detection} A \code{data.frame} containing taxa on the first column and the corresponding trais with NAs on the second column.
\item \code{duplicated_traits} If present, list the taxa with the same traits.
}
}
\description{
\Sexpr[results=rd, stage=render]{ lifecycle::badge("maturing") }

This function calculates the functional diversity.
}
\details{
Rao quadratic entropy (Q; Rao, 1982) is used to estimate Functional Diversity
because it has been considered more appropriate than other indices (Botta-Dukat,
2005; Ricotta, 2005). In this formula (Q) dij is the dissimilarity (ranging
from 0 to 1), between species i and j based on a set of specified functional
traits (i.e. effect traits, see below). This index is standardized by the maximum value to constrain the values
within the range of 0-1. Rao index is estimated using presence or abundance data
and the Euclidean transformed version of the traits-based Gower dissimilarity
matrix. For this, Gower's dissimilarity index (which ranges from 0 to 1) is used
because it can deal with traits of different nature and measuring scales
(continuous, nominal, binary, ordinal, etc.; see Podani 1999 for more information)

\deqn{Q = \sum_{i=1}^{S} \sum_{j=1}^{S} d_{ij} \ p_i \ p_j}

The \code{gower} distance refers to the mixed-variables coefficient of distance of Pavoine et al. (2009) as implemented in the \code{ade4} package.
This distance is meant to be used with fuzzy data.

Taxa without traits assigned in the trait dataset are removed from both the trait and abundance datasets.
}
\examples{
data(macro_ex)

data_bio <- as_biomonitor(macro_ex)
data_agr <- aggregate_taxa(data_bio)
data_ts <- assign_traits(data_agr)
# averaging
data_ts_av <- average_traits(data_ts)

col_blocks <- c(8, 7, 3, 9, 4, 3, 6, 2, 5, 3, 9, 8, 8, 5, 7, 5, 4, 4, 2, 3, 8)

f_divs(data_agr, trait_db = data_ts_av, type = "F", col_blocks = col_blocks)
f_divs(data_agr,
  trait_db = data_ts_av, type = "F", col_blocks = col_blocks,
  correction = "cailliez"
)

library(ade4)

rownames(data_ts_av) <- data_ts_av$Taxa
traits_prep <- prep.fuzzy(data_ts_av[, -1], col.blocks = col_blocks)

traits_dist <- ktab.list.df(list(traits_prep))
traits_dist <- dist.ktab(traits_dist, type = "F")

f_divs(data_agr, trait_db = traits_dist)
}
\references{
Biggs, R., Schluter, M., Biggs, D., Bohensky, E. L., BurnSilver, S.,
Cundill, G., ... & Leitch, A. M. (2012). Toward principles for enhancing the
resilience of ecosystem services. Annual Review of Environment and Resources,
37, 421-448.

Botta-Dukat, Z. (2005). Rao's quadratic entropy as a measure of
functional diversity based on multiple traits. Journal of Vegetation Science,
16(5), 533-540.

de Bello, F., Leps, J., Lavorel, S., & Moretti, M. (2007).
Importance of species abundance for assessment of trait composition:
an example based on pollinator communities. Community Ecology, 8(2), 163-170.

Elmqvist, T., Folke, C., Nystrom, M., Peterson, G.,
Bengtsson, J., Walker, B., & Norberg, J. (2003). Response diversity, ecosystem
change, and resilience. Frontiers in Ecology and the Environment, 1(9), 488-494.

Guillemot, N., Kulbicki, M., Chabanet, P., & Vigliola, L. (2011).
Functional redundancy patterns reveal non-random assembly rules in a
species-rich marine assemblage. PLoS One, 6(10), e26735.

Hevia, V., Martin-Lopez, B., Palomo, S., Garcia-Llorente, M., de Bello, F.,
& Gonzalez, J. A. (2017). Trait-based approaches to analyze links between the drivers
of change and ecosystem services: Synthesizing existing evidence and future challenges.
Ecology and evolution, 7(3), 831-844.

Hooper, D. U., Chapin, F. S., Ewel, J. J., Hector, A., Inchausti,
P., Lavorel, S., et al. (2005). Effects of biodiversity on ecosystem
functioning: a consensus of current knowledge. Ecological Monographs,
75(1), 3-35.

Lawton, J.H. & Brown, V.K. (1993) Redundancy in ecosystems.
Biodiversity and Ecosystem Function (eds E.-D. Schulze & H.A. Mooney),
pp. 255-270. Springer-Verlag, Berlin.

Pavoine, S., Vallet, J., Dufour, A. B., Gachet, S., & Daniel, H. (2009).
On the challenge of treating various types of variables: application for improving
the measurement of functional diversity. Oikos, 118(3), 391-402.

Pillar, V. D., Blanco, C. C., Muller, S. C., Sosinski, E. E.,
Joner, F., & Duarte, L. D. (2013). Functional redundancy and stability
in plant communities. Journal of Vegetation Science, 24(5), 963-974.

Podani, J. (1999). Extending Gower's general coefficient of
similarity to ordinal characters. Taxon, 331-340.

Rao, C. R. (1982). Diversity and dissimilarity coefficients:
a unified approach. Theoretical population biology, 21(1), 24-43.

Ricotta, C. (2005). A note on functional diversity measures.
Basic and Applied Ecology, 6(5), 479-486.

Rosenfeld, J. S. (2002). Functional redundancy in ecology
and conservation. Oikos, 98(1), 156-162.

Schmera, D., Heino, J., Podani, J., Eros, T., & Doledec, S. (2017).
Functional diversity: a review of methodology and current knowledge in
freshwater macroinvertebrate research. Hydrobiologia, 787(1), 27-44.

Walker, B. H. (1992). Biodiversity and ecological redundancy.
Conservation biology, 6(1), 18-23.
}
\seealso{
\link{aggregatoR}
}
