% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_eve.R
\name{f_eve}
\alias{f_eve}
\title{Functional evenness}
\usage{
f_eve(
  x,
  trait_db = NULL,
  tax_lev = "Taxa",
  type = NULL,
  traitSel = FALSE,
  col_blocks = NULL,
  nbdim = 2,
  distance = "gower",
  zerodist_rm = FALSE,
  correction = "none",
  traceB = FALSE,
  set_param = NULL
)
}
\arguments{
\item{x}{Result of \code{aggregate_taxa()}.}

\item{trait_db}{A trait dataset. Can be a \code{data.frame} ot a \code{dist} object.
Taxonomic level of the traits dataset must match those of the taxonomic dataset.
No automatic check is done by the \code{function}.}

\item{tax_lev}{Character string giving the taxonomic level used to retrieve
trait information. Possible levels are \code{Taxa}, \code{Species}, \code{Genus},
\code{Family} as returned by \code{aggregate_taxa()}.}

\item{type}{The type of variables speciefied in \code{trait_db}.
Must be one of \code{F}, fuzzy, or \code{C}, continuous.
If more control is needed please consider to provide \code{trait_db} as a \code{dist} object.
It works only when \code{trait_db} is a \code{data.frame}, otherwise ingored.}

\item{traitSel}{Select traits interactively.}

\item{col_blocks}{A vector that contains the number of modalities for each trait.
Not needed when \code{euclidean} distance is used.}

\item{nbdim}{Number of dimensions for the multidimensional functional spaces.
We suggest to keep \code{nbdim} as low as possible.
By default \code{biomonitoR} set the number of dimensions to 2. Select \code{auto} if you want the automated selection
approach according to Maire et al. (2015).}

\item{distance}{To be used to compute functional distances, \code{euclidean} or \code{gower}. Default to \code{gower}. See details.}

\item{zerodist_rm}{If \code{TRUE} aggregates taxa with the same traits.}

\item{correction}{Correction methods for negative eigenvalues, can be one of \code{none}, \code{lingoes}, \code{cailliez}, \code{sqrt} and \code{quasi}.
Ignored when type is set to \code{C}.}

\item{traceB}{If \code{TRUE} returns a list as specified in details.}

\item{set_param}{A list of parameters for fine tuning the calculations.
\code{max_nbdim} set the maximum number of dimension for evaluating the quality of the functional space.
\code{prec} can be \code{Qt} or \code{QJ}, please refere to the \code{convhulln} documentation for more information.
Deafault to \code{QJ}, less accurate but less prone to errors.
\code{tol} a tolerance threshold for zero, see the function \code{is.euclid}, \code{lingoes} and \code{cailliez} from the \code{ade4} for more details. Default to 1e-07.
If \code{cor.zero} is \code{TRUE}, zero distances are not modified. See functions \code{is.euclid}, \code{lingoes} and \code{cailliez} from the \code{ade4} for more details. Default to \code{TRUE}.

The \code{gower} distance refers to the mixed-variables coefficient of distance of Pavoine et al. (2009) as implemented in the \code{ade4} package.
This distance is meant to be used with fuzzy data.}
}
\value{
a vector with fuzzy functional richness results.
\enumerate{
\item \code{results} Results of \code{f_eve()}.
\item \code{traits} A \code{data.frame} containing the traits used for the calculations.
\item \code{taxa} A \code{data.frame} conaining the taxa used for the calculations.
\item \code{nbdim} Number of dimensions used after calculating the quality of functional spaces according to Maire et al. (2015).
\item \code{correction} The type of correction used.
\item \code{NA_detection} A \code{data.frame} containing taxa on the first column and the corresponding trais with NAs on the second column.
\item \code{duplicated_traits} If present, list the taxa with the same traits.
}
}
\description{
\Sexpr[results=rd, stage=render]{ lifecycle::badge("maturing") }

This function calculates functional evenness index.
}
\details{
Functional evenness represents a facet of functional diversity for a community with species distributed in a multidimensional functional space.
The metric is based on the minimum spanning tree which links all the species in the multidimensional functional space.
Then it quantifies the regularity with which species abundances are distributed along the spanning tree.

Functional evenness values are strictly positive and constrained between 0 and 1.
The higher they are, the higher the component of functional diversity they quantify is.
The measure quantifies the regularity with which the functional space is filled by species, weighted by their abundance and it is independent of species richness.
See formulas and more information in Mason et al. (2005) and Villeger et al. (2008).
}
\examples{
data(macro_ex)

data_bio <- as_biomonitor(macro_ex)
data_agr <- aggregate_taxa(data_bio)
data_ts <- assign_traits(data_agr)
# averaging
data_ts_av <- average_traits(data_ts)

col_blocks <- c(8, 7, 3, 9, 4, 3, 6, 2, 5, 3, 9, 8, 8, 5, 7, 5, 4, 4, 2, 3, 8)

f_eve(data_agr, trait_db = data_ts_av, type = "F", col_blocks = col_blocks)
f_eve(data_agr,
  trait_db = data_ts_av, type = "F", col_blocks = col_blocks,
  nbdim = 10, correction = "cailliez"
)

library(ade4)

rownames(data_ts_av) <- data_ts_av$Taxa
traits_prep <- prep.fuzzy(data_ts_av[, -1], col.blocks = col_blocks)

traits_dist <- ktab.list.df(list(traits_prep))
traits_dist <- dist.ktab(traits_dist, type = "F")

f_eve(data_agr, trait_db = traits_dist)
}
\references{
Mason, N. W., Mouillot, D., Lee, W. G., and
Wilson, J. B. (2005). Functional richness, functional evenness and functional
divergence: the primary components of functional diversity. Oikos, 111(1),
112-118.

Pavoine, S., Vallet, J., Dufour, A. B., Gachet, S., & Daniel, H. (2009).
On the challenge of treating various types of variables: application for improving
the measurement of functional diversity. Oikos, 118(3), 391-402.

Villeger, S., Mason, N. W., & Mouillot, D.
(2008). New multidimensional functional diversity indices for a
multifaceted framework in functional ecology. Ecology, 89(8), 2290-2301.
}
\seealso{
\link{aggregatoR}
}
