% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_red.R
\name{f_red}
\alias{f_red}
\title{Functional redundancy}
\usage{
f_red(
  x,
  trait_db = NULL,
  tax_lev = "Taxa",
  type = NULL,
  traitSel = FALSE,
  col_blocks = NULL,
  distance = "gower",
  zerodist_rm = FALSE,
  traceB = FALSE,
  correction = "none",
  set_param = NULL
)
}
\arguments{
\item{x}{results of function \code{aggregate_taxa()}.}

\item{trait_db}{a trait database. Can be a \code{data.frame} ot a \code{dist} object.
Taxonomic level of the traits database must match those of the taxonomic database.
No automatic check is done by the \code{function}.}

\item{tax_lev}{character string giving the taxonomic level used to retrieve
trait information. Possible levels are \code{"Taxa"}, \code{"Species"}, \code{"Genus"},
\code{"Family"} as returned by the \link{aggregatoR} function.}

\item{type}{the type of variables speciefied in \code{trait_db}.
Must be one of \code{F}, fuzzy, or \code{C}, continuous.
If more control is needed please consider to provide \code{trait_db} as a \code{dist} object.
It works only when \code{trait_db} is a \code{data.frame}, otherwise ingored.}

\item{traitSel}{interactively select traits.}

\item{col_blocks}{A vector that contains the number of modalities for each trait.
Not needed when \code{euclidean} distance is used.
By default \code{biomonitoR} select the optimal number of dimensions with the quality of the functional space approach.}

\item{distance}{to be used to compute functional distances, \code{euclidean} or \code{gower}. Default to \code{gower}.}

\item{zerodist_rm}{If \code{TRUE} aggregates taxa with the same traits.}

\item{traceB}{if \code{TRUE} ffrich will return a list as specified in details.}

\item{correction}{Correction methods for negative eigenvalues, can be one of \code{none}, \code{lingoes} , \code{cailliez}, \code{sqrt} and \code{quasi}.
Ignored when type is set to \code{C}.}

\item{set_param}{a list of parameters for fine tuning the calculations.
\code{tol} a tolerance threshold for zero, see the function \code{is.euclid}, \code{lingoes} and \code{cailliez} from the \code{ade4} for more details. Default to 1e-07.
\code{cor.zero} = \code{TRUE} if TRUE, zero distances are not modified. see the function \code{is.euclid}, \code{lingoes} and \code{cailliez} from the \code{ade4} for more details. Default to \code{TRUE}.}
}
\value{
a vector with fuzzy functional richness results.
\enumerate{
\item \strong{results}: results of \code{f_red()};
\item \strong{traits}: a data.frame containing the traits used for the calculations;
\item \strong{taxa}: a data.frame conaining the taxa used for th calculations;
\item \strong{nbdim}: number of dimensions used after calculatin the quality of functional spaces according to Maire et al., (2015);
\item \strong{correction}: the type of correction used.
\item \strong{NA_detection}: a data.frame containing taxa on the first column and the corresponding trais with NAs on the second column.
\item \strong{duplicated_traits}: if present, list the taxa with the same traits.
}
}
\description{
\Sexpr[results=rd, stage=render]{ lifecycle::badge("maturing") }

This function calculates the functional redundancy based on trait categories.
}
\details{
Functional redundancy (FR) is measured as the difference between taxonomic
diversity and functional diversity (de Bello et al., 2007). It relates positively
to ecosystem stability, resistance and resilience (Hooper et al. 2005; Guillemot
et al., 2011).

\deqn{ FR = D - Q}

The Gini-Simpson index is used to quantify Taxonomic Diversity (D, which ranges
from 0 to 1) where pi is the proportion of the abundance of taxa i in a
biological community.

\deqn{D = 1 - \sum_{i=1}^{S} p_i^{2}}

Rao quadratic entropy (Q; Rao, 1982) was used to estimate Functional Diversity
because it has been considered more appropriate than other indices (Botta-Dukat,
2005; Ricotta, 2005). In this formula (Q) dij is the dissimilarity (ranging
from 0 to 1), between species i and j based on a set of specified functional
traits (i.e. effect traits, see below). This index is standardized by the maximum value to constrain the values
within the range of 0-1. Rao index is estimated using presence or abundance data
and the Euclidean transformed version of the traits-based Gower dissimilarity
matrix. For this, Gower's dissimilarity index (which ranges from 0 to 1) is used
because it can deal with traits of different nature and measuring scales
(continuous, nominal, binary, ordinal, etc.; see Podani 1999 for more information)

\deqn{Q = \sum_{i=1}^{S} \sum_{j=1}^{S} d_{ij} \ p_i \ p_j}

Given that the concept of FR was originally developed to represent the number
of taxa contributing similarly to an ecosystem function (Walker 1992; Lawton and
Brown 1993; Rosenfeld, 2002), FR and therefore functional diversity should be
calculated using only effect traits. Effect traits are those biological
features that directly influence a specific function of the ecosystem
(e.g. productivity, nutrient cycling). See Schmera et al. (2017) and Hevia et al (2017) for more
information about effect traits in aquatic invertebrate communities. Regarding
the interpretation of the results, when taxa within a community differ
completely in their functional traits, then Q = D and thus FR = 0. On the
other hand, when all taxa have identical functional traits, then Q = 0 and
FR = D, and when in addition the number of taxa is very large and they are
equally abundant, then D (and in this case FR) approaches 1 (Pillar et al., 2013).
Although the concept of FR could suggest that functionally similar species may
compensate for the loss or failure of others, there is evidence that ecosystems
need such redundancy to perform their functions efficiently and stably over time
(Rosenfeld 2002; Biggs et al. 2012). In fact, a decrease in FR could be dramatic
in non-redundant communities since the loss or replacement of one species could
lead to loss of unique traits or functions (Hooper et al. 2005), increasing
ecosystem vulnerability (Elmqvist et al. 2003).

Taxa without traits assigned in the trait database are removed from both the trait and abundance databases.
}
\examples{
data(macro_ex)

data_bio <- as_biomonitor(macro_ex)
data_agr <- aggregate_taxa(data_bio)
data_ts <- assign_traits(data_agr)
# averaging
data_ts_av <- average_traits( data_ts )

col_blocks <- c( 8, 7, 3, 9, 4, 3, 6, 2, 5, 3, 9, 8, 8, 5, 7, 5, 4, 4, 2, 3, 8 )

f_red( data_agr , trait_db = data_ts_av , type = "F" , col_blocks = col_blocks )
f_red( data_agr , trait_db = data_ts_av , type = "F" , col_blocks = col_blocks ,
       correction = "cailliez" )

library( ade4 )

rownames( data_ts_av ) <- data_ts_av$Taxa
traits_prep <- prep.fuzzy( data_ts_av[ , -1 ], col.blocks = col_blocks )

traits_dist <- ktab.list.df( list( traits_prep ) )
traits_dist <- dist.ktab( traits_dist , type = "F" )

f_red( data_agr , trait_db = traits_dist   )



}
\references{
Biggs, R., Schluter, M., Biggs, D., Bohensky, E. L., BurnSilver, S.,
Cundill, G., ... & Leitch, A. M. (2012). Toward principles for enhancing the
resilience of ecosystem services. Annual Review of Environment and Resources,
37, 421-448.

Botta-Dukat, Z. (2005). Rao's quadratic entropy as a measure of
functional diversity based on multiple traits. Journal of Vegetation Science,
16(5), 533-540.

de Bello, F., Leps, J., Lavorel, S., & Moretti, M. (2007).
Importance of species abundance for assessment of trait composition:
an example based on pollinator communities. Community Ecology, 8(2), 163-170.

Elmqvist, T., Folke, C., Nystrom, M., Peterson, G.,
Bengtsson, J., Walker, B., & Norberg, J. (2003). Response diversity, ecosystem
change, and resilience. Frontiers in Ecology and the Environment, 1(9), 488-494.

Guillemot, N., Kulbicki, M., Chabanet, P., & Vigliola, L. (2011).
Functional redundancy patterns reveal non-random assembly rules in a
species-rich marine assemblage. PLoS One, 6(10), e26735.

Hevia, V., Martin-Lopez, B., Palomo, S., Garcia-Llorente, M., de Bello, F.,
& Gonzalez, J. A. (2017). Trait-based approaches to analyze links between the drivers
of change and ecosystem services: Synthesizing existing evidence and future challenges.
Ecology and evolution, 7(3), 831-844.

Hooper, D. U., Chapin, F. S., Ewel, J. J., Hector, A., Inchausti,
P., Lavorel, S., et al. (2005). Effects of biodiversity on ecosystem
functioning: a consensus of current knowledge. Ecological Monographs,
75(1), 3-35.

Lawton, J.H. & Brown, V.K. (1993) Redundancy in ecosystems.
Biodiversity and Ecosystem Function (eds E.-D. Schulze & H.A. Mooney),
pp. 255-270. Springer-Verlag, Berlin.

Pillar, V. D., Blanco, C. C., Muller, S. C., Sosinski, E. E.,
Joner, F., & Duarte, L. D. (2013). Functional redundancy and stability
in plant communities. Journal of Vegetation Science, 24(5), 963-974.

Podani, J. (1999). Extending Gower's general coefficient of
similarity to ordinal characters. Taxon, 331-340.

Rao, C. R. (1982). Diversity and dissimilarity coefficients:
a unified approach. Theoretical population biology, 21(1), 24-43.

Ricotta, C. (2005). A note on functional diversity measures.
Basic and Applied Ecology, 6(5), 479-486.

Rosenfeld, J. S. (2002). Functional redundancy in ecology
and conservation. Oikos, 98(1), 156-162.

Schmera, D., Heino, J., Podani, J., Eros, T., & Doledec, S. (2017).
Functional diversity: a review of methodology and current knowledge in
freshwater macroinvertebrate research. Hydrobiologia, 787(1), 27-44.

Walker, B. H. (1992). Biodiversity and ecological redundancy.
Conservation biology, 6(1), 18-23.
}
\seealso{
\link{aggregatoR}
}
