% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_rich.R
\name{f_rich}
\alias{f_rich}
\title{Functional richness}
\usage{
f_rich(
  x,
  traitDB = NULL,
  type = NULL,
  traitSel = FALSE,
  colB = NULL,
  nbdim = "auto",
  distance = "gower",
  zerodist_rm = FALSE,
  correction = "none",
  traceB = FALSE,
  set_param = NULL
)
}
\arguments{
\item{x}{results of function aggregatoR}

\item{traitDB}{a trait database. Can be a \code{data.frame} ot a \code{dist} object.
Taxonomic level of the traits database must match those of the taxonomic database.
No automatic check is done by the \code{function}.}

\item{type}{the type of variables speciefied in \code{traitDB}.
Must be one of \code{F}, fuzzy, or \code{C}, continuous.
If more control is needed please consider to provide \code{traitDB} as a \code{dist} object.
It works only when \code{traitDB} is a \code{data.frame}, otherwise ingored.}

\item{traitSel}{interactively select traits.}

\item{colB}{A vector that contains the number of modalities for each trait.
Not needed when \code{euclidean} distance is used.}

\item{nbdim}{number of dimensions for the multidimensional functional spaces.
We suggest to keep \code{nbdim} as low as possible.
By default \code{biomonitoR} select the optimal number of dimensions with the quality of the functional space approach.}

\item{distance}{to be used to compute functional distances, \code{euclidean} or \code{gower}. Default to \code{gower}.}

\item{zerodist_rm}{If \code{TRUE} aggregates taxa with the same traits.}

\item{correction}{Correction methods for negative eigenvalues, can be one of \code{none}, \code{lingoes} and \code{cailliez}.
Ignored when type is set to \code{C}.}

\item{traceB}{if \code{TRUE} ffrich will return a list as specified in details.}

\item{set_param}{a list of parameters for fine tuning the calculations.
\code{max_nbdim} set the maximum number of dimension for evaluating the quality of the functional space.
\code{prec} can be \code{Qt} or \code{QJ}, please refere to the \code{convhulln} documentation for more information.
Deafault to \code{QJ}, less accurate but less prone to errors.
\code{tol} a tolerance threshold for zero, see the function \code{is.euclid}, \code{lingoes} and \code{cailliez} from the \code{ade4} for more details. Default to 1e-07.
\code{cor.zero} = \code{TRUE} if TRUE, zero distances are not modified. see the function \code{is.euclid}, \code{lingoes} and \code{cailliez} from the \code{ade4} for more details. Default to \code{TRUE}.}
}
\value{
a vector with fuzzy functional richness results.
\enumerate{
\item \strong{results}: results of the ffred function;
\item \strong{traits}: a data.frame containing the traits used for the calculations;
\item \strong{taxa}: a data.frame conaining the taxa used for th calculations;
\item \strong{nbdim}: number of dimensions used after calculatin the quality of functional spaces according to Maire et al., (2015);
\item \strong{correction}: the type of correction used.
\item \strong{NA_detection}: a data.frame containing taxa on the first column and the corresponding trais with NAs on the second column.
\item \strong{duplicated_traits}: if present, list the taxa with the same traits.
\item \code{parent_child_pairs} For instance in Spanish \code{aspt} both Ferrissia and Planorbidae receive a score.
Abundances of the higher taxonomic level need therefore to be adjusted by subtracting the abundances of the lower taxonomic level.
}
}
\description{
This function calculates the functional richness based on trait categories.
}
\details{
Functional richness (FRic) represents the amount of functional space filled by
the community (Villeger et al., 2008) and it is related to the community use of
resources and productivity (Mason et al., 2005). FRic is defined by the trait
extremes and thus reflects the potential maximum functional dissimilarity.
FRic is calculated as the hypervolume enclosing the functional space filled
by the community. For this the convex hull approach is applied (Cornwell et al.,
2006) using the Quickhull algorithm (Barber et al., 1996) that estimates the minimum
convex hull which includes all the species considered in the previously defined
functional space. Basically, this algorithm determines the taxa in the most extreme
points of the functional space, links them to build the convex hull in order to
calculate the volume inside it. In particular, the convex hull of a set of points
S in n dimensions is the intersection of all convex sets containing S.
For N points , ..., , the convex hull C is then given by the expression:

\deqn{C = \sum_{j=1}^{N} \lambda_j \ p_j  : \ \lambda_j \geq \ for \ all \ j \ and \ \sum_{j=1}^{N} \lambda_j = 1}

The functional T dimensional space is built using a certain number of dimensions
(T) determined by the axes of a principal component analysis based on the trait
dissimilarity matrix. Using a poor quality functional space could led to
a biased assessment of FRic and false ecological conclusions so the number of
axes retained for its estimation is case-specific and decided following the
method proposed in Maire et al., (2015): a pragmatic approach consisting of
computing all the possible functional spaces and selecting the most
parsimonious one. This method uses the mSD index (mean squared deviation
between the initial functional distance and the scaled distance in the
functional space), which accounts explicitly for the deviation between
the initial and final distance and penalizes the strong deviation. The mSD
index has been widely used in statistics to assess errors and it has been
demonstrated to work in different contexts and situations (Maire et al., 2015).
In addition, when using Gower's distance the mSD ranges from 0 and 1,
which helps to interpret quality. Finally, the resulting FRic variable
is standardized by its maximum, ranging from 0 to 1. In addition, it must
be considered that the number of taxa must be higher than the number of traits
to have reliable FRic values (Villeger et al., 2008).
}
\examples{
data(macro_ex)

data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
data.ts <- traitScaling( data.agR )
# averaging
data.ts.av <- traitsMean( data.ts )

colB <- c( 8, 7, 3, 9, 4, 3, 6, 2, 5, 3, 9, 8, 8, 5, 7, 5, 4, 4, 2, 3, 8 )

f_rich( data.agR , traitDB = data.ts.av , type = "F" , colB = colB )
f_rich( data.agR , traitDB = data.ts.av , type = "F" , colB = colB ,
       nbdim = 10 , correction = "cailliez" )

library( ade4 )

rownames( data.ts.av ) <- data.ts.av$Taxa
traits.prep <- prep.fuzzy( data.ts.av[ , -1 ], col.blocks = colB )

traits.dist <- ktab.list.df( list( traits.prep ) )
traits.dist <- dist.ktab( traits.dist , type = "F" )

f_rich( data.agR , traitDB = traits.dist   )


}
\references{
Barber, C. B., Dobkin, D. P., & Huhdanpaa, H. (1996).
The quickhull algorithm for convex hulls. ACM Transactions on
Mathematical Software (TOMS), 22(4), 469-483.

Cornwell, W. K., Schwilk, D. W., & Ackerly, D. D. (2006).
A trait-based test for habitat filtering: convex hull volume.
Ecology, 87(6), 1465-1471

Maire, E., Grenouillet, G., Brosse, S., & Villeger, S. (2015).
How many dimensions are needed to accurately assess functional diversity?
A pragmatic approach for assessing the quality of functional spaces. Global
Ecology and Biogeography, 24(6), 728-740.

Mason, N. W., Mouillot, D., Lee, W. G., and
Wilson, J. B. (2005). Functional richness, functional evenness and functional
divergence: the primary components of functional diversity. Oikos, 111(1),
112-118.

Villeger, S., Mason, N. W., & Mouillot, D.
(2008). New multidimensional functional diversity indices for a
multifaceted framework in functional ecology. Ecology, 89(8), 2290-2301.
}
\seealso{
\link{aggregatoR}
}
