% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffdivs.R
\name{ffdivs}
\alias{ffdivs}
\title{Fuzzy coded functional diversity}
\usage{
ffdivs(
  x,
  traitDB = NULL,
  agg = FALSE,
  dfref = NULL,
  traitSel = FALSE,
  colB = NULL,
  taxLev = "Taxa",
  traceB = FALSE
)
}
\arguments{
\item{x}{results of function aggregatoR}

\item{traitDB}{a trait data base with a column \code{Taxa} and the other columns
containing the traits. Needed only if users use their own trait database.  If a trait has several modalities they should be
named as follow: TRAIT_MODALITY.

By default, the data base used is the one from Tachet \emph{et al} (2010) that
can be retrieved from
\href{https://www.freshwaterecology.info/}{freshwaterecology.info} website
(Schmidt-Kloiber & Hering, 2015).}

\item{agg}{should ffrich aggregate user's traitDB of higher taxonomic level? TRUE to aggregate, otherwise FALSE.
For instance, if user's traitDB has both Halesus and Limnephilidae, ffred will aggregate traits value if agg = TRUE.}

\item{dfref}{reference database to be used when a custom trait database is provided and agg = TRUE.
It should be the same reference database used in \link{asBiomonitor} when dfref = TRUE.}

\item{traitSel}{interactively select traits. See details for more info.}

\item{colB}{A vector that contains the number of modalities for each trait. Needed only if users use their own trait database.}

\item{taxLev}{character string giving the taxonomic level used to retrieve
trait information. Possible levels are \code{"Taxa"}, \code{"Species"}, \code{"Genus"},
\code{"Family"} as returned by the \link{aggregatoR} function.}

\item{traceB}{if TRUE ffred will return a list with 2 elements, the first being the ffrich values and the second the database used for the calculation. Useful to check missing taxa.}
}
\value{
a data.frame with 3 columns: Gini-Simpson richness, rao quadratic entropy and functional redundancy.\cr
If traceB is set to TRUE a list is provided with:
\enumerate{
\item \strong{results}: results of the ffred function;
\item \strong{traits}: a data.frame containing the traits used for the calculations;
\item \strong{taxa}: a data.frame conaining the taxa used for th calculations;
\item \strong{taxa_excluded}: a vector conaining the names of the taxa exluded from the calculations;
\item \strong{NA_detection}: a data.frame containing taxa on the first column and the corresponding trais with NAs on the second column.
}
}
\description{
TBD
}
\details{
Taxa without traits assigned in the trait database are removed from both the trait and abundance databases.
}
\note{
USE WITH CAUTION, STILL IN DEVELOPMENT.
}
\examples{
data(macro_ex)

data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
ffdivs(data.agR)

}
\references{
Biggs, R., Schluter, M., Biggs, D., Bohensky, E. L., BurnSilver, S.,
Cundill, G., ... & Leitch, A. M. (2012). Toward principles for enhancing the
resilience of ecosystem services. Annual Review of Environment and Resources,
37, 421-448.

Botta-Dukat, Z. (2005). Rao's quadratic entropy as a measure of
functional diversity based on multiple traits. Journal of Vegetation Science,
16(5), 533-540.

de Bello, F., Leps, J., Lavorel, S., & Moretti, M. (2007).
Importance of species abundance for assessment of trait composition:
an example based on pollinator communities. Community Ecology, 8(2), 163-170.

Elmqvist, T., Folke, C., Nystrom, M., Peterson, G.,
Bengtsson, J., Walker, B., & Norberg, J. (2003). Response diversity, ecosystem
change, and resilience. Frontiers in Ecology and the Environment, 1(9), 488-494.

Guillemot, N., Kulbicki, M., Chabanet, P., & Vigliola, L. (2011).
Functional redundancy patterns reveal non-random assembly rules in a
species-rich marine assemblage. PLoS One, 6(10), e26735.

Hevia, V., Martin-Lopez, B., Palomo, S., Garcia-Llorente, M., de Bello, F.,
& Gonzalez, J. A. (2017). Trait-based approaches to analyze links between the drivers
of change and ecosystem services: Synthesizing existing evidence and future challenges.
Ecology and evolution, 7(3), 831-844.

Hooper, D. U., Chapin, F. S., Ewel, J. J., Hector, A., Inchausti,
P., Lavorel, S., et al. (2005). Effects of biodiversity on ecosystem
functioning: a consensus of current knowledge. Ecological Monographs,
75(1), 3-35.

Lawton, J.H. & Brown, V.K. (1993) Redundancy in ecosystems.
Biodiversity and Ecosystem Function (eds E.-D. Schulze & H.A. Mooney),
pp. 255-270. Springer-Verlag, Berlin.

Pillar, V. D., Blanco, C. C., Muller, S. C., Sosinski, E. E.,
Joner, F., & Duarte, L. D. (2013). Functional redundancy and stability
in plant communities. Journal of Vegetation Science, 24(5), 963-974.

Podani, J. (1999). Extending Gower's general coefficient of
similarity to ordinal characters. Taxon, 331-340.

Rao, C. R. (1982). Diversity and dissimilarity coefficients:
a unified approach. Theoretical population biology, 21(1), 24-43.

Ricotta, C. (2005). A note on functional diversity measures.
Basic and Applied Ecology, 6(5), 479-486.

Rosenfeld, J. S. (2002). Functional redundancy in ecology
and conservation. Oikos, 98(1), 156-162.

Schmera, D., Heino, J., Podani, J., Eros, T., & Doledec, S. (2017).
Functional diversity: a review of methodology and current knowledge in
freshwater macroinvertebrate research. Hydrobiologia, 787(1), 27-44.

Walker, B. H. (1992). Biodiversity and ecological redundancy.
Conservation biology, 6(1), 18-23.
}
\seealso{
\link{aggregatoR}
}
