% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffred.R
\name{ffred}
\alias{ffred}
\title{Fuzzy coded functional redundancy}
\usage{
ffred(x, traitDB = NULL, agg = FALSE, dfref = NULL, traitSel = FALSE,
  colB = NULL, taxLev = "Family", traceB = FALSE)
}
\arguments{
\item{x}{results of function aggregatoR}

\item{traitDB}{a trait data base with a column \code{Taxa} and the other columns
containing the traits. If a trait has several modalities they should be
named as follow: TRAIT_MODALITY.

By default, the data base used is the one from Tachet \emph{et al} (2010) that
can be retrieved from
\href{https://www.freshwaterecology.info/}{freshwaterecology.info} website
(Schmidt-Kloiber & Hering, 2015).}

\item{agg}{should ffrich aggregate user's traitDB of higher taxonomic level? TRUE to aggregate, otherwise FALSE.
For instance, if user's traitDB has both Halesus and Limnephilidae, ffrich will aggregate traits value if ADD = TRUE.}

\item{dfref}{reference database to be used when a custom trait database is provided and agg equals to TRUE.}

\item{traitSel}{interactively select traits.}

\item{colB}{A vector that contains the number of modalities for each trait}

\item{taxLev}{character string giving the taxonomic level used to retrieve
trait information. Possible levels are \code{"Taxa"}, \code{"Species"}, \code{"Genus"},
\code{"Family"} as returned by the \link{aggregatoR} function.}

\item{traceB}{if TRUE ffrich will return a list with 2 elements, the first being the ffrich values and the second the database used for the calculation. Useful to check missing taxa.}
}
\value{
a data.frame with 3 columns: Gini-Simpson richness, rao quadratic entropy and functional redundancy.\cr
If traceB is set to TRUE a list is provided with:
\enumerate{
\item \emph{results}: results of the ffred function;
\item \emph{traits}: a data.frame containing the traits used for the calculations;
\item \emph{taxa}: a data.frame conaining the taxa used for th calculations;
\item \emph{taxa_not_used}: a vector conaining the names of the taxa exluded from the calculations. Taxa are excluded from the calculations if they have NA values at least in one of the trait modalities;
\item \emph{problematic traits}: traits containing NAs for \emph{taxa_not_used}
\item \emph{traits_not_used}: trait modalities excluded from the calculations because they have 0 value for each species.
}
}
\description{
This function calculates the functional richness based on trait categories.
}
\details{
Functional redundancy (FR) is measured as the difference between taxonomic
diversity and functional diversity (de Bello et al., 2007). It relates positively
to ecosystem stability, resistance and resilience (Hooper et al. 2005; Guillemot
et al., 2011).

\deqn{ FR = D - Q}

The Gini-Simpson index is used to quantify Taxonomic Diversity (D, which ranges
from 0 to 1) where pi is the proportion of the abundance of taxa i in a
biological community.

\deqn{D = 1 - \sum_{i=1}^{S} p_i^{2}}

Rao quadratic entropy (Q; Rao, 1982) was used to estimate Functional Diversity
because it has been considered more appropriate than other indices (Botta-Dukat,
2005; Ricotta, 2005). In this formula (Q) dij is the dissimilarity (ranging
from 0 to 1), between species i and j based on a set of specified functional
traits. This index is standardized by the maximum value to constrain the values
within the range of 0-1. Rao index is estimated using presence or abundance data
and the Euclidean transformed version of the traits-based Gower dissimilarity
matrix. For this, Gower's dissimilarity index (which ranges from 0 to 1) is used
because it can deal with traits of different nature and measuring scales
(continuous, nominal, binary, ordinal, etc.; see Podani 1999 for more information)

\deqn{Q = \sum_{i=1}^{S} \sum_{j=1}^{S} d_{ij} \ p_i \ p_j}

Given that the concept of FR was originally developed to represent the number
of taxa contributing similarly to an ecosystem function (Walker 1992; Lawton and
Brown 1993; Rosenfeld, 2002), FR and therefore functional diversity should be
calculated using only effect traits. Effect traits are those biological
features that directly influence a specific function of the ecosystem
(e.g. productivity, nutrient cycling). See Schmera et al., 2017 for more
information about effect traits in aquatic invertebrate communities. Regarding
the interpretation of the results, when taxa within a community differ
completely in their functional traits, then Q = D and thus FR = 0. On the
other hand, when all taxa have identical functional traits, then Q = 0 and
FR = D, and when in addition the number of taxa is very large and they are
equally abundant, then D (and in this case FR) approaches 1 (Pillar et al., 2013).
Although the concept of FR could suggest that functionally similar species may
compensate for the loss or failure of others, there is evidence that ecosystems
need such redundancy to perform their functions efficiently and stably over time
(Rosenfeld 2002; Biggs et al. 2012). In fact, a decrease in FR could be dramatic
in non-redundant communities since the loss or replacement of one species could
lead to loss of unique traits or functions (Hooper et al. 2005), increasing
ecosystem vulnerability (Elmqvist et al. 2003).

Taxa with no traits are removed from both the trait and abundance databases.
}
\note{
USE WITH CAUTION, STILL IN DEVELOPMENT.
}
\examples{
data(macro_ex)

data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
ffred(data.agR)

}
\references{
Biggs, R., Schluter, M., Biggs, D., Bohensky, E. L., BurnSilver, S.,
Cundill, G., ... & Leitch, A. M. (2012). Toward principles for enhancing the
resilience of ecosystem services. Annual Review of Environment and Resources,
37, 421-448.

Botta-Dukat, Z. (2005). Rao's quadratic entropy as a measure of
functional diversity based on multiple traits. Journal of Vegetation Science,
16(5), 533-540.

de Bello, F., Leps, J., Lavorel, S., & Moretti, M. (2007).
Importance of species abundance for assessment of trait composition:
an example based on pollinator communities. Community Ecology, 8(2), 163-170.

Elmqvist, T., Folke, C., Nystrom, M., Peterson, G.,
Bengtsson, J., Walker, B., & Norberg, J. (2003). Response diversity, ecosystem
change, and resilience. Frontiers in Ecology and the Environment, 1(9), 488-494.

Guillemot, N., Kulbicki, M., Chabanet, P., & Vigliola, L. (2011).
Functional redundancy patterns reveal non-random assembly rules in a
species-rich marine assemblage. PLoS One, 6(10), e26735.

Hooper, D. U., Chapin, F. S., Ewel, J. J., Hector, A., Inchausti,
P., Lavorel, S., et al. (2005). Effects of biodiversity on ecosystem
functioning: a consensus of current knowledge. Ecological Monographs,
75(1), 3-35.

Lawton, J.H. & Brown, V.K. (1993) Redundancy in ecosystems.
Biodiversity and Ecosystem Function (eds E.-D. Schulze & H.A. Mooney),
pp. 255-270. Springer-Verlag, Berlin.

Pillar, V. D., Blanco, C. C., Muller, S. C., Sosinski, E. E.,
Joner, F., & Duarte, L. D. (2013). Functional redundancy and stability
in plant communities. Journal of Vegetation Science, 24(5), 963-974.

Podani, J. (1999). Extending Gower's general coefficient of
similarity to ordinal characters. Taxon, 331-340.

Rao, C. R. (1982). Diversity and dissimilarity coefficients:
a unified approach. Theoretical population biology, 21(1), 24-43.

Ricotta, C. (2005). A note on functional diversity measures.
Basic and Applied Ecology, 6(5), 479-486.

Rosenfeld, J. S. (2002). Functional redundancy in ecology
and conservation. Oikos, 98(1), 156-162.

Schmera, D., Heino, J., Podani, J., Eros, T., & Doledec, S. (2017).
Functional diversity: a review of methodology and current knowledge in
freshwater macroinvertebrate research. Hydrobiologia, 787(1), 27-44.

Walker, B. H. (1992). Biodiversity and ecological redundancy.
Conservation biology, 6(1), 18-23.
}
\seealso{
\link{aggregatoR}
}
