% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffrich.R
\name{ffrich}
\alias{ffrich}
\title{Fuzzy coded functional richness}
\usage{
ffrich(x, traitDB = NULL, agg = FALSE, dfref = NULL,
  traitSel = FALSE, colB = NULL, taxLev = "Family", traceB = FALSE,
  nbdim = 7, metric = "Gower", corr_method = FALSE)
}
\arguments{
\item{x}{results of function aggregatoR}

\item{traitDB}{a trait data base with a column \code{Taxa} and the other columns
containing the traits. If a trait has several modalities they should be
named as follow: TRAIT_MODALITY.

By default, the data base used is the one from Tachet \emph{et al} (2010) that
can be retrieved from
\href{https://www.freshwaterecology.info/}{freshwaterecology.info} website
(Schmidt-Kloiber & Hering, 2015).}

\item{agg}{should ffrich aggregate user's traitDB of higher taxonomic level? TRUE to aggregate, otherwise FALSE.
For instance, if user's traitDB has both Halesus and Limnephilidae, ffrich will aggregate traits value if ADD = TRUE.}

\item{dfref}{reference database to be used when a custom trait database is provided and agg equals to TRUE.}

\item{traitSel}{interactively select traits.}

\item{colB}{A vector that contains the number of modalities for each trait}

\item{taxLev}{character string giving the taxonomic level used to retrieve
trait information. Possible levels are \code{"Taxa"}, \code{"Species"}, \code{"Genus"},
\code{"Family"} as returned by the \link{aggregatoR} function.}

\item{traceB}{if TRUE ffrich will return a list with 2 elements, the first being the ffrich values and the second the database used for the calculation. Useful to check missing taxa.}

\item{nbdim}{maximum number of dimensions for multidimensional functional spaces. By default, nbdim=7                   ##
Final number of dimensions depends on the number of positive eigenvalues (after correction) obtained with PCoA}

\item{metric}{to be used to compute functional distance, "Euclidean" or "Gower" (=default)}

\item{corr_method}{TRUE will aplly the Cailliez correction.}
}
\value{
a vector with fuzzy functional richness results.
\enumerate{
\item \strong{results}: results of the ffred function;
\item \strong{traits}: a data.frame containing the traits used for the calculations;
\item \strong{taxa}: a data.frame conaining the taxa used for th calculations;
\item \strong{nbdim}: number of dimensions used after calculatin the quality of functional spaces according to Maire et al., (2015);
\item \strong{taxa_not_used}: a vector conaining the names of the taxa exluded from the calculations. Taxa are excluded from the calculations if they have NA values at least in one of the trait modalities;
\item \strong{problematic traits}: traits containing NAs for \emph{taxa_not_used}
\item \strong{traits_not_used}: trait modalities excluded from the calculations because they have 0 value for each species.
}
}
\description{
This function calculates the functional richness based on trait categories.
}
\details{
Functional richness (FRic) represents the amount of functional space filled by
the community (Villeger et al., 2008) and it is related to the community use of
resources and productivity (Mason et al., 2005). FRic is defined by the trait
extremes and thus reflects the potential maximum functional dissimilarity.
FRic is calculated as the hypervolume enclosing the functional space filled
by the community. For this the convex hull approach is applied (Cornwell et al.,
2006) using the Quickhull algorithm (Barber et al., 1996) that estimates the minimum
convex hull which includes all the species considered in the previously defined
functional space. Basically, this algorithm determines the taxa in the most extreme
points of the functional space, links them to build the convex hull in order to
calculate the volume inside it. In particular, the convex hull of a set of points
S in n dimensions is the intersection of all convex sets containing S.
For N points , ..., , the convex hull C is then given by the expression:

\deqn{C = \sum_{j=1}^{N} \lambda_j \ p_j  : \ \lambda_j \geq \ for \ all \ j \ and \ \sum_{j=1}^{N} \lambda_j = 1}

The functional T dimensional space is built using a certain number of dimensions
(T) determined by the axes of a principal component analysis based on the trait
dissimilarity matrix. Using a poor quality functional space could led to
a biased assessment of FRic and false ecological conclusions so the number of
axes retained for its estimation is case-specific and decided following the
method proposed in Maire et al., (2015): a pragmatic approach consisting of
computing all the possible functional spaces and selecting the most
parsimonious one. This method uses the mSD index (mean squared deviation
between the initial functional distance and the scaled distance in the
functional space), which accounts explicitly for the deviation between
the initial and final distance and penalizes the strong deviation. The mSD
index has been widely used in statistics to assess errors and it has been
demonstrated to work in different contexts and situations (Maire et al., 2015).
In addition, when using Gower's distance the mSD ranges from 0 and 1,
which helps to interpret quality. Finally, the resulting FRic variable
is standardized by its maximum, ranging from 0 to 1. In addition, it must
be considered that the number of taxa must be higher than the number of traits
to have reliable FRic values (Villeger et al., 2008).
}
\note{
USE WITH CAUTION, STILL IN DEVELOPMENT.
}
\examples{
data(macro_ex)

data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
ffrich(data.agR, taxLev = "Taxa")

}
\references{
Barber, C. B., Dobkin, D. P., & Huhdanpaa, H. (1996).
The quickhull algorithm for convex hulls. ACM Transactions on
Mathematical Software (TOMS), 22(4), 469-483.

Cornwell, W. K., Schwilk, D. W., & Ackerly, D. D. (2006).
A trait-based test for habitat filtering: convex hull volume.
Ecology, 87(6), 1465-1471

Maire, E., Grenouillet, G., Brosse, S., & Villeger, S. (2015).
How many dimensions are needed to accurately assess functional diversity?
A pragmatic approach for assessing the quality of functional spaces. Global
Ecology and Biogeography, 24(6), 728-740.

Mason, N. W., Mouillot, D., Lee, W. G., and
Wilson, J. B. (2005). Functional richness, functional evenness and functional
divergence: the primary components of functional diversity. Oikos, 111(1),
112-118.

Villeger, S., Mason, N. W., & Mouillot, D.
(2008). New multidimensional functional diversity indices for a
multifaceted framework in functional ecology. Ecology, 89(8), 2290-2301.
}
\seealso{
\link{aggregatoR}
}
