% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_t.R
\name{flow_t}
\alias{flow_t}
\title{Flow-T index}
\usage{
flow_t(x, tax_lev = "Taxa", type = "ab", trans = NULL)
}
\arguments{
\item{x}{Result of \code{aggregate_taxa()}.}

\item{tax_lev}{Taxonomic level at which the calculation has to be performed.}

\item{type}{Presence only \code{po} or abundance \code{ab}.}

\item{trans}{The function used to transform the abundances, by default \code{log1p}.}
}
\description{
This function calculates the Flow-T index.
}
\details{
This function calculates the Flow_T index developed by Laini et al. (2022).
Flow-T is calculated using the current velocity preference categories for each sample using:

\deqn{Flow-T= \frac{\sum_{i=1}^{n}\log(A_i + 1)(m_i + f_i)}{\sum_{i=1}^{n}\log(A_i + 1)}}

Where n represents the number of taxa in a sample, A_i the abundance of the ith taxon, mi and fi the
"moderate" and "fast" velocity preference classes of the ith taxon according to Tachet et al. (2010).
The logarithm of the abundance can be replaced by 1 to obtain the presence-absence version of the Flow-T
index.
}
\examples{

data(mi_prin)
data_bio <- as_biomonitor(mi_prin)
data_agr <- aggregate_taxa(data_bio)
data_ts <- assign_traits(data_agr)

}
\references{
Laini, A., Burgazzi, G., Chadd, R., England, J., Tziortzis, I., Ventrucci, M., Vezza, P., Viaroli, P., Wood, P.J. & Guareschi, S. (2022). Using invertebrate functional traits to improve flow variability assessment within European rivers. Science of The Total Environment, 832, 155047.

Tachet, H., Richoux, P., Bournaud, M., Usseglio-Polatera, P., (2010). Invertebres d'eau douce: Systematique, biologie, ecologie, edition revue et augmentee. CNRS EDITIONS, Paris.
}
\seealso{
\link{as_biomonitor} \link{assign_traits}
}
