% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iucn_taxa.R
\name{get_iucn_taxa_tree}
\alias{get_iucn_taxa_tree}
\title{Create biomonitoR reference dataset from GBIF}
\usage{
get_iucn_taxa_tree(x, ref_from_tree = FALSE, token = NULL)
}
\arguments{
\item{x}{Vector containing your taxa list.}

\item{ref_from_tree}{Create a reference database in the \code{biomonitoR} format. See \link{ref_from_tree}.}

\item{token}{See IUCN API token. Detailed information at https://apiv3.iucnredlist.org/api/v3/token}
}
\description{
Function to create a reference biomonitoR dataset using API from the Global Biodiversity Information Facility (GBIF).
}
\examples{
\dontrun{

data(macro_ex)
refDB <- get_iucn_taxa_tree(macro_ex$Taxa)
refDB_bior <- refDB$taxonomy
data_asb <- as_biomonitor(macro_ex, dfref = refDB_bior, overwrite = TRUE)
data_agR <- aggregate_taxa(data_asb)

}
}
\seealso{
\link{as_biomonitor} \link{get_gbif_taxa_tree} \link{get_nbn_taxa_tree} \link{get_worms_taxa_tree}
}
