% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nbn_taxa.R
\name{get_nbn_taxa_tree}
\alias{get_nbn_taxa_tree}
\title{Create \code{biomonitoR} reference dataset from the National Biodiversity Network (NBN)}
\usage{
get_nbn_taxa_tree(x, ref_from_tree = FALSE)
}
\arguments{
\item{x}{Vector contain your Taxa list. This is the base list to create your custom reference database.}

\item{ref_from_tree}{Create a reference database in the \code{biomonitoR} format. See \link{ref_from_tree}.}
}
\description{
Function to create a reference \code{biomonitoR} dataset using API from the National Biodiversity Network (NBN).
}
\details{
Check carefully the results. In the NBN database, sometimes the subgenus information
is embedded in the species name (e.g Genus (Subgenus) species). This can cause problems to other \code{biomonitoR} functions.
}
\examples{

\dontrun{

data(macro_ex)
dfref_c <- get_gbif_taxa_tree(macro_ex[, "Taxa"])
data_asb <- as_biomonitor(macro_ex, dfref = dfref_nbn$taxonomy)
data_agg <- aggregate_taxa(data_asb)

}
}
\seealso{
\link{as_biomonitor} \link{get_gbif_taxa_tree} \link{get_iucn_taxa_tree} \link{get_worms_taxa_tree}
}
