% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_db.R
\name{get_reference_db}
\alias{get_reference_db}
\title{Function to download reference taxonomic trees, traits dataset,
and biomonitoring index scores.}
\usage{
get_reference_db(db_name = "mi_cuba_2023")
}
\arguments{
\item{db_name}{dataset name to download. See details to know the available data sets.}
}
\description{
This function allows the user to download several data sets useful for \code{biomonitoR} analyses.
See details to discover which data sets you can download.
}
\details{
Available data set:
\enumerate{
\item mi_cuba: Taxonomic data set of macroinvertebrates from Cuba (Torres-Cambas et al., 2023).
}
}
\examples{
db_reference <- get_reference_db(db_name = "mi_cuba_2023")

}
\keyword{data}
\keyword{download}
\keyword{reference}
\keyword{sets}
