% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxa_abundance.R
\name{get_taxa_abundance}
\alias{get_taxa_abundance}
\title{get_taxa_abundance}
\usage{
get_taxa_abundance(x, taxa = NULL, rel = FALSE)
}
\arguments{
\item{x}{result of the function aggregatoR.}

\item{taxa}{a taxon or a vector of taxa.}

\item{rel}{if TRUE calculates relative abundance. Default to FALSE.}
}
\description{
This function calculates the absolute or relative abundance of a taxon or of a set taxa.
}
\details{
This function does not check for parent-child pairs. For instance if the vector of taxa contains both the genus \emph{Baetis} and the
family Baetidae, the abundance of the genus \emph{Baetis} will be double counted. Check the function \code{\link{ambiguousSolver}} for a solution
to this problem.
}
\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
get_taxa_abundance(data.agR, taxa = "Ephemeroptera")
get_taxa_abundance(data.agR, taxa = c("Setodes", "Orthocladiinae"), rel = TRUE)
}
\seealso{
\code{\link{aggregatoR}} \code{\link{ambiguousSolver}}
}
\keyword{aggregatoR}
