% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxa_richness.R
\name{get_taxa_richness}
\alias{get_taxa_richness}
\title{get_taxa_richness}
\usage{
get_taxa_richness(x, taxa = NULL, tax_lev = NULL)
}
\arguments{
\item{x}{Result of the function \code{aggregate_taxa()}.}

\item{taxa}{a taxon or a vector of taxa.}

\item{tax_lev}{taxonomic level at which the richness has to be calculated. It could be also a vector of taxonomic levels.}
}
\description{
This function calculates the absolute richness of a taxon or of a set taxa at the taxonomic level provided by the user.
}
\examples{
data(macro_ex)
data_bio <- as_biomonitor(macro_ex)
data_agR <- aggregate_taxa(data_bio)
get_taxa_richness(data_agR, taxa = "Ephemeroptera", tax_lev = "Family")
}
\seealso{
\code{\link{aggregate_taxa}}
}
\keyword{aggregate_taxa}
