% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_worms_taxa.R
\name{get_worms_taxa_tree}
\alias{get_worms_taxa_tree}
\title{Function for creating reference databases using API from World Register of Marine Species (WORMS)}
\usage{
get_worms_taxa_tree(x, ref_from_tree = FALSE)
}
\arguments{
\item{x}{Vector contain your Taxa list.}

\item{ref_from_tree}{Create a reference database in the \code{biomonitoR} format. See \link{ref_from_tree}.}
}
\description{
This function allows creating our custom reference database using API from World Register of Marine Species (WORMS).
}
\examples{
\dontrun{

data(macro_ex)
dfref_worms <- get_worms_taxa_tree(macro_ex[, "Taxa"])
data_asb <- as_biomonitor(macro_ex, dfref = dfref_worms$taxonomy)
data_agg <- aggregate_taxa(data_asb)

}
}
\seealso{
\link{as_biomonitor}
}
