% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold.R
\name{gold}
\alias{gold}
\title{gold}
\usage{
gold(x)
}
\arguments{
\item{x}{results of function aggregatoR}

\item{taxLev}{the taxonomic level for calculating EPT richness.}
}
\description{
This function calculates the 1 - GOLD metric.
}
\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
gold(data.agR)
}
\seealso{
\code{\link{aggregatoR}}
}
\keyword{ept}
