% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibmr.R
\name{ibmr}
\alias{ibmr}
\title{Biologique Macrofitique en Riviere index}
\usage{
ibmr(
  x,
  method = "ita",
  exceptions = NULL,
  coverage_coef = FALSE,
  coverage_classes = c(0.1, 1, 10, 50),
  traceB = FALSE
)
}
\arguments{
\item{x}{Result of \code{aggregate_taxa()}.}

\item{method}{The implementation of IBMR. the default method is \code{ita} (Minciardi et al., 2009).
Users can provide their own data.frame (see examples) with a column called \emph{Taxon} and the column of scores called \emph{Scores}.}

\item{exceptions}{Taxa that need to be excluded from the calculation.
This option can be useful, for instance, to exclude an alien species belonging to an autochthonous family.}

\item{coverage_coef}{Are data already represented with coverage coefficients? Set to \code{TRUE} if this is not the case.}

\item{coverage_classes}{Thresholds to transform abundances into coverage classes.}

\item{traceB}{If set to \code{TRUE} a list as specified below will be returned.}
}
\value{
If \code{traceB} is set to \code{TRUE} a list with the following elements will be returned:
\itemize{
\item \code{results} Result of \code{aspt()}.
\item \code{taxa_df} The \code{data.frame} used for the calculation containing the abundance of taxa receiving a score.
\item \code{exceptions} A \code{data.frame} containing the changes made by excluding the taxa included in \code{exceptions}.
\item \code{parent_child_pairs} For instance in Spanish ASPT both \emph{Ferrissia} and Planorbidae receive a score.
Abundances of the higher taxonomic level need therefore to be adjusted by subtracting the abundances of the lower taxonomic level.
}
}
\description{
\Sexpr[results=rd, stage=render]{ lifecycle::badge("experimental") }

This function calculates the IBMR index following Minciardi et al. (2009).
}
\details{
The function \code{ibmr()} calculates the index as follow:
\deqn{ibmr= \frac{\sum_{i=1}^{S} C{i} E{i} K{i}} {\sum_{i=1}^{S} E{i} K{i}}}

where S is the number of taxa, i the ith taxon, Ci the sensitivity score, Ei the stenoecia coefficient
and Ki the coverage coefficient.

Be careful when using data already transformed in coverage classes. The function \code{ibmr()} will sum
the classes if needed and this will result in wrong results. For example, if a taxon has only genus level
scores and two species of this genus are present, \code{ibmr()} will sum their coverage classes. This can result
in taxa having a coverage class higher than the highest coverage class allowed.

All the information used for \code{ibmr()} calculation can be retrieved with the function \link{show_scores}.
}
\examples{
data(oglio)
oglio_asb <- as_biomonitor(oglio, group = "mf", FUN = bin)
oglio_agg <- aggregate_taxa(oglio_asb)
ibmr(oglio_agg)
}
\references{
Minciardi, M. R., Spada, C. D., Rossi, G. L., Angius, R., Orru, G., Mancini, L., Pace, G., Marcheggiani, S., Puccinelli, C. (2009). Metodo per la valutazione e la classificazione dei corsi d'acqua utilizzando la comunita delle macrofite acquatiche. ENEA Rapporto Tecnico RT/2009/23/ENEA.
}
\seealso{
\link{aggregate_taxa}
}
\keyword{ibmr}
