% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indNumb.R
\name{indNumb}
\alias{indNumb}
\title{indNumb}
\usage{
indNumb(d, method = "a")
}
\arguments{
\item{d}{results of function aggregatoR}

\item{method}{the only option is "b" (see details).}
}
\description{
Function for calculating the number of Taxa according to taxalist of desired index.
}
\details{
Only implemented for the Italian version of the STAR_ICMi
}
\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
indNumb(data.agR)
}
\seealso{
\code{\link{aggregatoR}}, \code{\link{speNumb}}, \code{\link{genNumb}}, \code{\link{famNumb}}
}
\keyword{indNumb}
