% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_ex.R
\docType{data}
\name{macro_ex}
\alias{macro_ex}
\title{Apennine invertebrate dataset}
\format{A data frame with 34 rows and 3 variables.}
\usage{
macro_ex
}
\description{
This dataset is derived from samples taken during the NOACQUA project aiming at evaluating "Community responses and ecosystem processes in
intermittent streams".
It consists of information taken at different taxonomic levels, mainly at genus and family levels.
Macroinvertebrates were collected at ten points per site with a surber net with a frame of 0.05 m2 and a mesh net of 500 micrometers.
All the available microhabitats were investigated. Data are reported as abundances.
The NOACQUA project (Prot. 201572HW8F) was funded by the Italian Ministry of Education, University and Research.
}
\keyword{datasets}
