% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageTraits.R
\name{manageTraits}
\alias{manageTraits}
\title{manageTraits}
\usage{
manageTraits(x, method = "nearest+-", traceB = FALSE)
}
\arguments{
\item{x}{result of the \code{traitScaling} function.}

\item{method}{can be \code{nearest}, \verb{nearest+}, \verb{nearest-}, \verb{nearest+-} and \verb{neareast-+}.
Please see details for further information.}

\item{traceB}{if \code{TRUE} it will return a vector containing taxa excluded from the selection process
because they did not meet the reuirments of the selection.}
}
\description{
A function to select traits based on taxonomic distance.
}
\details{
Method \code{nearest} select the traits belonging to the nearest taxa irrispective of their position
in the taxonomic tree.
Method \verb{nearest+} select the traits belonging to the nearest taxa that have a taxonomic level equal or finer
than the target one. Method \verb{nearest-} do the opposite.
Method \verb{nearest+} select the traits belonging to the nearest taxa giving priority to taxa having
taxonomic level equal or finer than the target one. Method \verb{nearest-+} do the opposite.
}
\examples{
data(macro_ex)

data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
data.ts <- traitScaling( data.agR )

# select only the nearest traits
data.ts.sub <- manageTraits( data.ts , method = "nearest+-" )

}
