% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oglio.R
\docType{data}
\name{oglio}
\alias{oglio}
\title{Macrophytes example dataset}
\format{A data frame with 67 rows and 7 variables.}
\usage{
data(oglio)
}
\description{
These data were collected in the Oglio river, northern Italy. This data.frame consists of cover abundances on a scale from 0 to 100.
}
\keyword{datasets}
