% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcoaQuality.R
\name{pcoaQuality}
\alias{pcoaQuality}
\title{pcoaQuality}
\usage{
pcoaQuality(x, type = "none", method = NULL, tresh = 0.7, nbdim = 15)
}
\arguments{
\item{x}{distance matrix}

\item{type}{correction. Default to \code{none}, other choices are \code{cailliez}, \code{lingoes} and \code{sqrt}.}

\item{method}{three evaluation methods are availabe: \code{cor}, \code{legendre} and \code{maire}.}

\item{tresh}{treshold used for the \code{cor} method.}

\item{maximum}{number of dimensions for the \code{maire} method. Keep it as small as possible.}
}
\description{
This function merge the quality of the functional space.
}
