% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biomonitoR.R
\name{plot.biomonitoR}
\alias{plot.biomonitoR}
\title{Plot biomonitoR object}
\usage{
\method{plot}{biomonitoR}(
  x,
  tax_lev = "Taxa",
  method = "ward.D2",
  relative = TRUE,
  unassigned = FALSE,
  trans = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class biomonitoR generated with \code{aggregate_taxa}.}

\item{tax_lev}{Taxonomic level to be plotted.}

\item{method}{Clustering method.}

\item{relative}{Plot absolute or relative abundance.}

\item{unassigned}{Remove unassigned taxa.}

\item{trans}{Transformation to be applied to abundance data.}

\item{...}{Further arguments to be passed to \code{plot}.}
}
\description{
barplot based on clustering methods
}
\examples{

data(macro_ex)
data_bio <- as_biomonitor(mi_prin)
data_agr <- aggregate_taxa(data_bio)

}
