% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_indicator_taxa.R
\name{plot_indicator_taxa}
\alias{plot_indicator_taxa}
\title{Plot indicator taxa}
\usage{
plot_indicator_taxa(
  x,
  clusters = NULL,
  tax_lev = "Taxa",
  nmax = 10,
  thresh = 25,
  method = "ward.D2",
  parent = "Order",
  opacity = 0.8,
  color = NULL
)
}
\arguments{
\item{x}{Result of \code{aggregate_taxa()}.}

\item{clusters}{Sites clusters. If \code{NULL} an automatic clustering procedure will be used. Otherwise, provide a
\code{data.frame} with 2 columns, the first called \code{sites} where store sites name while the second called \code{clusters} where
store cluster groups. Please avoid groups with numeric values.}

\item{tax_lev}{Taxonomic level to be plotted.}

\item{nmax}{Maximum number of clusters when \code{clusters} is \code{NULL}.}

\item{thresh}{Threshold to identify indicator taxa, ranging from 0 to 100.}

\item{method}{Clustering method if \code{clusters} is \code{NULL}.}

\item{parent}{Parent taxonomic level. It must be coarser than \code{tax_lev}.}

\item{opacity}{Set the opacity of the color link.}

\item{color}{Set the color link.}
}
\description{
A sankey plot of indicator taxa.
}
\details{
Indicator Taxa Analysis (IndVal; Dufrene and Legendre, 1997) is a method to find indicator species and species assemblages characterizing groups of sites.
The function \code{plot_indicator_taxa} implements the group-equalized IndVal for both presence-absence and abundance data (De Caceres and Legendre, 2009). Please consider
that \code{plot_indicator_taxa} is intended to visualize the main trends present in the data. For having more control on the
IndVal calculation consider the \code{indicspecies} package (De Caceres and Legendre, 2009).
}
\examples{

data(macro_ex)
data_bio <- as_biomonitor(mi_prin)
data_agr <- aggregate_taxa(data_bio)
plot_indicator_taxa(data_agr)
}
\references{
Dufrene, M., & Legendre, P. (1997). Species assemblages and indicator species: the need for a flexible asymmetrical approach. Ecological monographs, 67(3), 345-366.

De Caceres, M., & Legendre, P. (2009). Associations between species and groups of sites: indices and statistical inference. Ecology, 90(12), 3566-3574.
}
