% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_rename.R
\name{quick_rename}
\alias{quick_rename}
\title{quick_rename}
\usage{
quick_rename(x, group = "mi", write_table = FALSE)
}
\arguments{
\item{x}{a data.frame as specified in details}

\item{group}{biotic group of interest. Possible values are \code{mi} for macroinvertebrates and \code{mf} for macrophytes. Default to \code{mi}.}

\item{write_table}{if TRUE \code{quick_rename()} will save a txt file with changes provided by the user}
}
\description{
This function allow the user to change the wrong taxa names according to the deault reference databases implemented in \code{biomonitoR}.
}
\details{
data.frame must have a column called "Taxa" where put species, genus or family names. See data(macro_ex) for an example dataset.\cr
The function \code{quick_rename()} will suggest correct name while allowing the user to insert a name (Enter taxon name).
}
\examples{
data(macro_ex)
macro_ex_mod <- quick_rename(macro_ex, group = "mi")
}
\seealso{
\code{\link{asBiomonitor}}
}
\keyword{as_biomonitor}
