% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_taxa.R
\name{remove_taxa}
\alias{remove_taxa}
\title{remove_taxa}
\usage{
remove_taxa(x, taxa = NULL)
}
\arguments{
\item{x}{result of the function aggregatoR.}

\item{taxa}{a taxon or a vector of taxa.}
}
\description{
Remove a taxon or a set of taxa and create a new \code{data.frame}.
}
\details{
This function does not check for parent-child pairs but it should handle them pretty well.
}
\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
remove_taxa(data.agR, taxa = "Ephemeroptera")
}
\seealso{
\code{\link{aggregatoR}} \code{\link{ambiguousSolver}}
}
\keyword{aggregatoR}
