% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename}
\alias{rename}
\title{rename}
\usage{
rename(x, write.table = F)
}
\arguments{
\item{x}{a data.frame as specified in details}
}
\description{
This function allow the user to change the wron taxa names.
}
\details{
data.frame must have a column called "Taxa" where put species, genus or family names. See data(macro_ex) for an example dataset.\cr
The function rename will suggest correct name and allow the user to insert a name (Enter taxon name).
}
\examples{
data(macro_ex)
macro_ex.mod <- rename(macro_ex)
}
\seealso{
\code{\link{asBiomonitor}}
}
\keyword{asBiomonitor}
