% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectPcoaAxes.R
\name{selectPcoaAxes}
\alias{selectPcoaAxes}
\title{selectPcoaAxes}
\usage{
selectPcoaAxes(x, method = "legendre", tresh = NULL, nbdim = 15)
}
\arguments{
\item{x}{a distance matrix.}

\item{method}{the measure of the quality of the functional space. Can be \code{cor}, \code{legendre} and \code{maire}.}

\item{tresh}{the treshold for selecting the axis. By default is 0.7 for \code{cor} and \code{legendre} and
0.01 for \code{maire}.}

\item{nbdim}{the maximum number of dimension when the option is set to \code{maire}. Default to 15.}
}
\description{
\Sexpr[results=rd, stage=render]{ lifecycle::badge("deprecated") }

Select the number of the axis for the calculation of functional richness.
}
\details{
This function provide some information to select the number of axis for the calculation of
functional richness. As implemented in biomonitoR, functional richness require a pcoa to be computed from
the trait distances. For calculating functional richness a subset of the number of axis originating
from the pcoa is kept. The choiche of the number of axis can affect the calculation and some
information about how much the subset represents the original distance matrix. To this purpose, 3 measures
are provide. The \code{cor} option represents the correlation between the distance of the points in the
reduced space and distance matrix. The distance of points in the reduced space is calculated
with the \code{dist} function on the coordinates of the selected number of axis.
A p-value based on a mantel statistic is also provided.
The \code{legendre} option calculates the R^2 like ratio as described by Legendre and Legendre (2008)
and implemented in the FD package.
The \code{maire} option calculate the quality of the functional space according to Maire et al. (2015).
The function selectPcoaAxis provide also information about the euclidean property of the trait
distance. The lack of this property can lead to biases and some solutions are proposed here and
described in Legendre and Legendre (2008). Please take a look also to the help of the FD
function dbFD for further information.
}
\references{
Maire, E., Grenouillet, G., Brosse, S., & Villeger, S. (2015).
How many dimensions are needed to accurately assess functional diversity?
A pragmatic approach for assessing the quality of functional spaces. Global
Ecology and Biogeography, 24(6), 728-740.

Legendre, P. and L. Legendre (1998) Numerical Ecology.
2nd English edition. Amsterdam: Elsevier.
}
\seealso{
\code{\link{aggregatoR}}
}
\keyword{aggregatoR}
