% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon.R
\name{shannon}
\alias{shannon}
\alias{simpson}
\alias{margalef}
\alias{menhinick}
\alias{pielou}
\alias{berpar}
\alias{brill}
\alias{esimpson}
\alias{invberpar}
\alias{invsimpson}
\alias{mcintosh}
\alias{odum}
\title{shannon}
\usage{
shannon(x, base = 2, taxLev = "Family")
}
\arguments{
\item{x}{results of function aggregatoR}

\item{base}{the base of the logarithm}

\item{taxLev}{taxonimc level on which the calculation has to be made.}
}
\description{
Functions for calculating shannon, simpson, margalef, menhinick, pielou and other indices.
}
\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
shannon(data.agR)

# natural logarithm
shannon(data.agR, base=exp(1))
}
\seealso{
\code{\link{aggregatoR}}
}
\keyword{margalef,}
\keyword{menhinick,}
\keyword{pielou}
\keyword{shannon,}
\keyword{simpson,}
