% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_scores.R
\name{show_scores}
\alias{show_scores}
\title{show_scores}
\usage{
show_scores(index = NULL, method = NULL)
}
\arguments{
\item{index}{name of the index for which information are needed.}

\item{method}{method of the index specified in \code{index}.}
}
\description{
This function print the databases used for the calculation of several indices implemented in biomonitoR.
}
\examples{
show_scores(index = "aspt", method = "spa")
show_scores(index = "life", method = "extence")
}
\seealso{
\code{\link{aspt}}, \code{\link{bmwp}}, \code{\link{life}}, \code{\link{whpt}}, \code{\link{psi}}, \code{\link{epsi}}
}
\keyword{asBiomonitor}
