% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speNumb.R
\name{speNumb}
\alias{speNumb}
\title{speNumb}
\usage{
speNumb(x)
}
\arguments{
\item{x}{results of function aggregatoR}
}
\description{
Functions for calculating species, genus, family and order Richness
}
\details{
By now only species, genus and family richness calculation are reliable. This is because order assignment for order in the reference database is not completely covered. Unassigned taxon are exluded from the calculations.
}
\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
genNumb(data.agR)
}
\seealso{
\code{\link{aggregatoR}}
}
\keyword{famNumb,}
\keyword{genNumb,}
\keyword{ordNumb}
\keyword{speNumb,}
