% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_frame_asb.R
\name{subset.asb}
\alias{subset.asb}
\title{subset.asb}
\usage{
\method{subset}{asb}(x, ..., taxa = NULL, exclude = NULL)
}
\arguments{
\item{x}{Results of function \code{as_biomonitor()}.}

\item{...}{Rurther arguments to be passed to or from other methods.}

\item{taxa}{A taxon or a vector of taxa to retain.}

\item{exclude}{A taxon or a vector of taxa to exclude. For example, this option is useful to exclude alien taxa.}
}
\description{
\code{summary} method for class \code{asb}.
\Sexpr[results=rd, stage=render]{ lifecycle::badge("experimental") }
}
\examples{
data(macro_ex)
data_bio <- as_biomonitor(macro_ex)

# select EPT Taxa (Ephemeroptera, Plecoptera and Trichoptera)

subset(data_bio, taxa = c("Ephemeroptera", "Plecoptera", "Trichoptera"))

# select Trichoptera excluding the trichopteran family Hydropsychidae

tricho <- subset(data_bio, taxa = "Trichoptera", exclude = "Hydropsychidae")
tricho_agg <- aggregate_taxa(tricho)

# exclude Chironomidae
subset(data_bio, exclude = "Chironomidae")
}
