% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_bin.R
\name{to_bin}
\alias{to_bin}
\title{Abundance to presence-absence}
\usage{
to_bin(x)
}
\arguments{
\item{x}{A \code{data.frame}}
}
\description{
This function transforms abundance to presence-absence.
}
\details{
\code{to_bin} will transform values greter than 0 to 1.
}
\examples{

data(macro_ex)

to_bin(macro_ex)

}
