% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tree.R
\docType{data}
\name{Tree}
\alias{Tree}
\title{Taxonomic tree example}
\format{A data frame with 5 rows and 6 variables:
\describe{
\item{Phylum}{Phylum level}
\item{Class}{Class level}
\item{Order}{Order level}
\item{Family}{Family level}
\item{Genus}{Genus level}
\item{Species}{Species level}
}}
\usage{
Tree
}
\description{
This is an example taxonomic tree to build reference databases suitable for biomonitoR.
}
\keyword{datasets}
