% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{abundance}
\alias{abundance}
\title{abundance}
\usage{
abundance(x, tax_lev = "Taxa", unassigned = FALSE)
}
\arguments{
\item{x}{Result of the function \code{aggregate_taxa()}.}

\item{tax_lev}{Taxonomic level.}

\item{unassigned}{Does unassigned taxa need to be taken into account? If yes set unassigned to \code{TRUE}, otherwise \code{FALSE}.}
}
\description{
Calculates taxa abundances at the desired taxonomic level.
Unassigned taxa can be included or excluded from the calculation.
}
\details{
If unassigned is set to \code{TRUE} abundances are equal among taxonomic levels.
}
\examples{
data(macro_ex)
data_bio <- as_biomonitor(macro_ex)
data_agr <- aggregate_taxa(data_bio)
abundance(data_agr, tax_lev = "Family")
abundance(data_agr, tax_lev = "Family", unassigned = TRUE)
}
\seealso{
\link{aggregate_taxa} \link{get_taxa_abundance}
}
\keyword{abundance}
