% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abuTax.R
\name{abuTax}
\alias{abuTax}
\title{abuTax}
\usage{
abuTax(x, taxa = NULL, rel = FALSE)
}
\arguments{
\item{x}{results of function aggregatoR.}

\item{taxa}{a Taxon or a vector of taxa.}

\item{rel}{if TRUE calculates relative abundance. default =F.}
}
\description{
This function calculates the absolute or relative abundance of a Taxon or of a set Taxa.
}
\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
abuTax(data.agr, taxa = "Ephemeroptera")
abuTax(data.agr, taxa = c("Plecopotera", "Ephemeroptera"))
}
\seealso{
\code{\link{aggregatoR}}
}
\keyword{aggregatoR}
