% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregatoR.R
\name{aggregatoR}
\alias{aggregatoR}
\title{aggregatoR}
\usage{
aggregatoR(z)
}
\arguments{
\item{x}{results of function asBiomonitoR}
}
\description{
This function prepares data for further calculations.
}
\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
}
\seealso{
\code{\link{asBiomonitor}}
}
\keyword{aggregatoR}
