% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregatoR.R
\name{aggregatoR}
\alias{aggregatoR}
\title{aggregatoR}
\usage{
aggregatoR(x)
}
\arguments{
\item{x}{results of function asBiomonitoR}
}
\description{
This function prepares data for further calculations.
}
\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)

# example with macrophytes
data(oglio)

oglio.asb <- asBiomonitor(oglio, group = "mf")
oglio.agg <- aggregatoR(oglio.asb)
speNumb(oglio.agg)
genNumb(oglio.agg)
famNumb(oglio.agg)
}
\seealso{
\code{\link{asBiomonitor}}
}
\keyword{aggregatoR}
