% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregatoR.R
\name{aggregatoR}
\alias{aggregatoR}
\title{aggregatoR}
\usage{
aggregatoR(x, FUN = sum)
}
\arguments{
\item{x}{results of function asBiomonitor}

\item{FUN}{the function to be applied for aggregating to higher taxonomic levels.
Must be sum for both abundances and presence-absence data.
Default to \code{sum}. Change if you know what you are doing.}
}
\description{
This function prepares data for further calculations.
}
\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)

# example for macrophytes
data(oglio)

oglio.asb <- asBiomonitor(oglio, group = "mf")
oglio.agg <- aggregatoR(oglio.asb)
richness(oglio.agg , taxLev = "Species")
richness(oglio.agg , taxLev = "Genus")
richness(oglio.agg , taxLev = "Family")
}
\seealso{
\code{\link{asBiomonitor}}
}
\keyword{aggregatoR}
