% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allindices.R, R/berpar.R, R/brill.R,
%   R/esimpson.R, R/invberpar.R, R/invsimpson.R, R/margalef.R, R/mcintosh.R,
%   R/menhinick.R, R/odum.R, R/pielou.R, R/shannon.R, R/simpson.R
\name{allindices}
\alias{allindices}
\alias{shannon}
\alias{simpson}
\alias{margalef}
\alias{menhinick}
\alias{pielou}
\alias{berpar}
\alias{brill}
\alias{esimpson}
\alias{invberpar}
\alias{invsimpson}
\alias{mcintosh}
\alias{odum}
\alias{berpar}
\alias{brill}
\alias{esimpson}
\alias{invberpar}
\alias{invsimpson}
\alias{margalef}
\alias{mcintosh}
\alias{menhinick}
\alias{odum}
\alias{pielou}
\alias{shannon}
\alias{simpson}
\title{Functions for calculating diversity, eveness and dominance indices}
\usage{
allindices(x, taxLev = "Family", base = 2)

berpar(x, taxLev = "Family")

brill(x, taxLev = "Family")

esimpson(x, taxLev = "Family")

invberpar(x, taxLev = "Family")

invsimpson(x, taxLev = "Family")

margalef(x, taxLev = "Family")

mcintosh(x, taxLev = "Family")

menhinick(x, taxLev = "Family")

odum(x, base = 2, taxLev = "Family")

pielou(x, base = 2, taxLev = "Family")

shannon(x, base = 2, taxLev = "Family")

simpson(x, taxLev = "Family")
}
\arguments{
\item{x}{results of function aggregatoR}

\item{taxLev}{taxonimc level on which the calculation has to be made.}

\item{base}{the base of the logarithm}
}
\description{
Functions for calculating shannon, simpson, margalef, menhinick, pielou and other indices.
}
\section{Functions}{
\itemize{
\item \code{berpar}: Berger-Parker dominance index

\item \code{brill}: Brillouin index

\item \code{esimpson}: Simpson's Index

\item \code{invberpar}: inverse Berger-Parker

\item \code{invsimpson}: Inverse Simpson

\item \code{margalef}: Margalef diversity index

\item \code{mcintosh}: McIntosh dominance index

\item \code{menhinick}: Menhinick index

\item \code{odum}: Odum index

\item \code{pielou}: Pielou's evenness index

\item \code{shannon}: Brillouin index

\item \code{simpson}: Simpson's Index of Diversity
}}

\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
allindices(data.agR)
shannon(data.agR)

# natural logarithm
shannon(data.agR, base=exp(1))
}
\seealso{
\code{\link{aggregatoR}}
}
\keyword{margalef,}
\keyword{menhinick,}
\keyword{pielou}
\keyword{shannon,}
\keyword{simpson,}
