% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allindices.R, R/berpar.R, R/brill.R,
%   R/esimpson.R, R/fisher.R, R/invberpar.R, R/invsimpson.R, R/margalef.R,
%   R/mcintosh.R, R/menhinick.R, R/pielou.R, R/shannon.R, R/simpson.R
\name{allindices}
\alias{allindices}
\alias{shannon}
\alias{simpson}
\alias{margalef}
\alias{menhinick}
\alias{pielou}
\alias{berpar}
\alias{brill}
\alias{esimpson}
\alias{invberpar}
\alias{invsimpson}
\alias{mcintosh}
\alias{fisher}
\title{Functions for calculating diversity, eveness and dominance indices.}
\usage{
allindices(x, tax_lev = "Taxa", base = exp(1))

berpar(x, tax_lev = "Taxa")

brill(x, tax_lev = "Taxa")

esimpson(x, tax_lev = "Taxa")

fisher(x, tax_lev = "Taxa")

invberpar(x, tax_lev = "Taxa")

invsimpson(x, tax_lev = "Taxa")

margalef(x, tax_lev = "Taxa")

mcintosh(x, tax_lev = "Taxa")

menhinick(x, tax_lev = "Taxa")

pielou(x, base = exp(1), tax_lev = "Taxa")

shannon(x, base = exp(1), tax_lev = "Taxa")

simpson(x, tax_lev = "Taxa")
}
\arguments{
\item{x}{Result of the function \code{aggregate_taxa()}.}

\item{tax_lev}{Taxonomic level on which the calculation has to be made.}

\item{base}{The base of the logarithm}
}
\description{
Functions for calculating shannon, simpson, margalef, menhinick, pielou and other indices.
}
\details{
Shannon index:
\deqn{H'=-\sum_{i=1}^{S} p_{i}\log(p_{i})}
Pielou index:
\deqn{J'=\frac{H'}{\log(S)}}
Margalef diversity:
\deqn{D_{mg}=\frac{(S - 1)}{\log(N)}}
Menhinick diversity:
\deqn{D_{mn}=\frac{S}{\sqrt(N)}}
Brillouin index:
\deqn{HB=\frac{\log(N) -\sum(\log(n_{i})) }{N}}
Simpson's index is calculated as D. biomonitoR returns 1 - D and 1/D when \code{index} is set to \code{Simpson} or \code{Invsimpson}.
\deqn{D=\sum_{i=1}^{S} p_{i}^{2}}
Simpson's evenness:
\deqn{D_{1/D}=\frac{1/D}{\sqrt(S)}}
Berger-Parker index:
\deqn{d=\frac{N_{max}}{N}}
The inverse of this index is also provided when \code{index} is set to \code{Invberpar}. McIntosh's diversity:
\deqn{D=\frac{N-U}{N-\sqrt(N)}}
where
\deqn{U=\sqrt(\sum_{i=1}^{S} n_{i}^{2})}
Fisher alpha is calculated as follow:
\deqn{\alpha=\frac{N(1-x)}{x}}
where x is estimated from the iterative solution of:
\deqn{\frac{S}{N}=-\frac{N(1-x)}{x} \log(1-x)}

p_i is the proportion of individuals found in the i-th species, n_i is the number of individuals found in the i-th species, S the species richness,
N the number of individuals and N_max the number of individuals in the most abundant species. All the indices are calculated according to Magurran (2004).
}
\section{Functions}{
\itemize{
\item \code{berpar}: Berger-Parker index

\item \code{brill}: Brillouin index

\item \code{esimpson}: Simpson's evenness

\item \code{fisher}: Fisher alpha

\item \code{invberpar}: Inverse Berger-Parker

\item \code{invsimpson}: Inverse Simpson

\item \code{margalef}: Margalef diversity index

\item \code{mcintosh}: McIntosh dominance index

\item \code{menhinick}: Menhinick index

\item \code{pielou}: Pielou's evenness index

\item \code{shannon}: Shannon index

\item \code{simpson}: Simpson's Index of Diversity
}}

\examples{
data(macro_ex)
data_bio <- as_biomonitor(macro_ex)
data_agr <- aggregate_taxa(data_bio)
allindices(data_agr)
shannon(data_agr)

# base 2
shannon(data_agr, base = 2)
}
\references{
Magurran, A. E. (2004). Measuring biological diversity. Blackwell Science ltd.
}
\seealso{
\link{aggregate_taxa}
}
\keyword{margalef,}
\keyword{menhinick,}
\keyword{pielou}
\keyword{shannon,}
\keyword{simpson,}
