% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allrich.R, R/richness.R
\name{allrich}
\alias{allrich}
\alias{richness}
\title{Function for calculating richness at different taxonomic levels}
\usage{
allrich(x)

richness(x, taxLev = "Family")
}
\arguments{
\item{x}{results of function aggregatoR}

\item{taxLev}{taxonomic level on which the calculation has to be made.}
}
\description{
This function allows the calculation of richness at different taxonomic levels and of the total abundance.
}
\section{Functions}{
\itemize{
\item \code{richness}: Taxonomic richness
}}

\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
richness(data.agR, taxLev = "Family" )
}
\seealso{
\code{\link{aggregatoR}}
}
\keyword{allrich,}
\keyword{richness}
