% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiguousSolver.R
\name{ambiguousSolver}
\alias{ambiguousSolver}
\title{ambiguousSolver}
\usage{
ambiguousSolver(x, method = "higher")
}
\arguments{
\item{x}{result of function aggregatoR.}

\item{method}{only method \emph{higher} is currently implemented. See details.}
}
\description{
Solve the ambiguous assignment of taxa.
}
\details{
Taxonomic dataset often contains ambiguous taxa due to the difficulites in identify damaged and juvenile organisms. For instance Chironominae and Chironomidae could reported as in the same sample. Several techinques has been
proposed to solve this issue (see Cuffney et al., 2007). Currently ambiguousSolver has only one algorithm that allow to assing the abundances of lowr taxonomic level to those of higher taxonomic level.
}
\examples{
data( macro_ex )
data.bio <- asBiomonitor( macro_ex )
data.agR <- aggregatoR( data.bio )
df <- ambiguousSolver( data.agR )
df.bio <- asBiomonitor( df )
df.agR <- aggregatoR( df.bio )
}
\references{
Cuffney, T. F., Bilger, M. D., & Haigler, A. M. (2007). Ambiguous taxa: effects on the characterization and interpretation of invertebrate assemblages. Journal of the North American Benthological Society, 26(2), 286-307.
}
\seealso{
\code{\link{aggregatoR}}
}
\keyword{aggregatoR}
