% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asBiomonitor.R
\name{asBiomonitor}
\alias{asBiomonitor}
\title{asBiomonitor - deprecated}
\usage{
asBiomonitor(
  x,
  group = "mi",
  dfref = NULL,
  to_change = "default",
  FUN = sum,
  correct_names = FALSE,
  traceB = FALSE
)
}
\arguments{
\item{x}{a data.frame with a column called "Taxa" where store taxa names and samples on the other columns (see the example macro_ex).}

\item{group}{biotic group of interest. Possible values are \code{mi} for macroinvertebrates, \code{mf} for macrophytes and \code{fi} for fish. The choice will set the right reference database for the specified group.
This option will not be considered if a custom reference database is provided. Default to \code{mi}.}

\item{dfref}{a custom reference database that replaces the reference database.}

\item{to_change}{a \code{data.frame} specifying the taxa name that needs to be changed.
This \code{data.frame} needs a column called \emph{Taxon} containing the taxon to aggregate and a column called \emph{Correct_Taxon} with the aggregation specifications.
By default, when group is set to \code{mi} Hydracarina, Hydracnidia and Acariformes are changed to Trombidiformes.}

\item{FUN}{the function to be applied for aggregating rows with duplicated taxa names.
It should be \code{sum} for abundances, while it should be \code{bin} for presence-absence data. Default to \code{sum}.}

\item{correct_names}{if \code{TRUE} alternative names will be suggested for taxa not found in the reference database. Default to
\code{FALSE}, with which the unrecognized taxa will be removed.}

\item{traceB}{track changes in taxa names.}
}
\description{
\Sexpr[results=rd, stage=render]{ lifecycle::badge("deprecated") }

This function merge the user dataframe with a reference database and suggest corrections for mispelled names.
}
\details{
The function \code{asBiomonitor} checks the taxonomy of the data.frame provided by the user and suggests correction for mispelled names.
If one or more taxa names of \code{x} are not present in the reference database or the spell checker is not able to find any suggestion the user is asked to exit.
This behaviour is to assure consistency with other functions implemented in biomonitoR.
Default references databases are provided for macroinvertebrates and macrophytes.
Both databases heavily rely on the information provided by the \href{https://www.freshwaterecology.info/}{freshwaterecology.info} website.
If \code{dfref} is not NULL a custom dictionary will be saved in the working directory to let the \code{asBiomonitor} function work correctly.
If you are unable to build a reference database by your own please check the function \code{\link{refFromTree}} for a possible solution.
\code{asBiomonitor}  returns an object of class \code{biomonitoR} togheter with one of the classes \code{abundance} or \code{bin}.
The function \code{\link{quickRename}} works as the \code{asBiomonitor} but returns
a data.frame without the biomonitoR format.
\code{asBiomonitor} aggregates all the rows with the same name with the option \code{FUN} and converts all the \code{NA} to 0.
If only 1 and 0 are present \code{x} will be imported as presence-absence.
When \code{group = mi} Hydracarina, Hydracnidia or Acariformes are changed to Trombidiformes given the uncertain taxonomic status of this group.
}
\references{
Schmidt-Kloiber, A., & Hering, D. (2015). www.freshwaterecology.info -
An online tool that unifies, standardises and codifies more than
20,000 European freshwater organisms and their ecological preferences.
Ecological indicators, 53, 271-282.
}
\seealso{
\code{\link{quickRename}} \code{\link{refFromTree}} \code{\link{quickRename}}
}
\keyword{asBiomonitor}
