% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asBiomonitor.R
\name{asBiomonitor}
\alias{asBiomonitor}
\title{asBiomonitor}
\usage{
asBiomonitor(x, group = "mi", dfref = NULL, overwrite = F)
}
\arguments{
\item{x}{a data.frame with a column called "Taxa" where put taxa names (species, genus, family, etc) and samples on the other columns (see the example data macro_ex).}

\item{group}{biotic group of interest. Possible values are "mi" for macroinvertebrates and "mf" for macrophytes. This option will not be considered if overwrite is set to FALSE in \code{dfref}.}

\item{dfref}{allow the user to improve (if overwrite = F) or to replace (if overwrite = T) the reference database.}

\item{overwrite}{if set to T the reference database is replaced with the one provided by the user.}
}
\description{
This function merge the user dataframe with the reference database (be retrieved from
\href{https://www.freshwaterecology.info/}{freshwaterecology.info} website
(Schmidt-Kloiber & Hering, 2015)). Options to improve or replace the reference database are provided.
}
\details{
\code{asBiomonitor} checks the correctness of taxa names in the data.frame provided by the user. If names are correct the function will process the data.frame to a biomonitor object, otherwise it will provide suggestions for correct names. If dfref = T a custom dictionary will be saved in the working directory.
When using the default database the taxa names provided by the user need to be consistent with the taxonomy of \href{https://www.freshwaterecology.info/}{freshwaterecology.info}, otherwise the user is asked to exit. This behaviour is to assure consistency with other functions implemented in biomonitoR.
}
\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex, group = "mi")
}
\references{
Schmidt-Kloiber, A., & Hering, D. (2015). www. freshwaterecology.
info-An online tool that unifies, standardises and codifies more than
20,000 European freshwater organisms and their ecological preferences.
Ecological indicators, 53, 271-282.
}
\seealso{
\code{\link{quickRename}}
}
\keyword{asBiomonitor}
