% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspt.R
\name{aspt}
\alias{aspt}
\title{aggregatoR}
\usage{
aspt(d, method = "a")
}
\arguments{
\item{method}{a,b or i. See details.}

\item{x}{results of function aggregatoR}
}
\description{
Functions for calculating BMWP and ASPT
}
\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
aspt(data.agR)
}
\seealso{
\code{\link{aggregatoR}}
}
\keyword{aggregatoR}
