% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspt.R
\name{aspt}
\alias{aspt}
\title{aspt}
\usage{
aspt(d, method = "a")
}
\arguments{
\item{d}{results of function aggregatoR}

\item{method}{the formulation of BMWP needed to calculate ASPT. Possible choises are "a" (Armitage et al. 1983), "b" (Davy-Bowker et al. 2007) and "i" (MAGRAMA 2011)}
}
\description{
This function calculates the Average Score Per Taxon following Armitage et al. (1983), Davy-Bowker et al. (2007) and Alba-Tercedor & Sanchez-Ortega (1988) implementations.
}
\details{
ASPT represents the average scores of the families that receive the score. Armitage scores are not reliable yet, since taxonomy has to be revised (e.g. Elminthidae are present instead of Elmidae). Davy-Bowker implementation take into account composite taxa as follow:
\enumerate{
  \item Psychomyiidae (inc. Ecnomidae)
  \item Rhyachopilidae (inc. Glossomatidae)
  \item Limnephilidae (inc. Apatanidae)
  \item Ancylidae (inc. Acroloxidae)
  \item Gammaridae (inc. Crangonyctidae & Niphargidae)
  \item Hydrophilidae (inc. Hydraenidae, Helophoridae)
  \item Tipulidae (inc. Limoniidae, Pediciidae & Cylindrotomidae)
  \item Planariidae (inc. Dugesidae)
  \item Hydrobiidae (inc. Bithyniidae)
  \item Oligochaeta (all the families)
}
}
\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
aspt(data.agR)
aspt(data.agR, method = "i")
}
\references{
Armitage, P. D., Moss, D., Wright, J. F., & Furse, M. T. (1983). The performance of a new biological water quality score system based on macroinvertebrates over a wide range of unpolluted running-water sites. Water research, 17(3), 333-347.

Davy-Bowker J., Clarke R., Corbin T., Vincent H, Pretty J., Hawczak A., Blackburn J., Murphy J., Jones I., 2008. River Invertebrate Classification Tool. Final report. WFD72C. SNIFFER. 276 pp

MAGRAMA-Ministerio de Agricultura y medio Ambiente (2011) Protocolo de muestreo y laboratorio de fauna bentónica de invertebrados en ríos vadeables. ML-Rv-I-2011, Cód, 23 pp.
}
\seealso{
\code{\link{aggregatoR}}
}
\keyword{aggregatoR}
