% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_traits.R, R/average_traits.R,
%   R/sample_traits.R
\name{assign_traits}
\alias{assign_traits}
\alias{average_traits}
\alias{sample_traits}
\title{Assign traits to taxa at different taxonomic levels}
\usage{
assign_traits(
  x,
  trait_db = NULL,
  group = "mi",
  tax_lev = "Taxa",
  dfref = NULL,
  filter_by_distance = NULL
)

average_traits(x, col_blocks = NULL)

sample_traits(x)
}
\arguments{
\item{x}{Results of \code{aggregate_taxa()}.}

\item{trait_db}{A trait \code{data.frame} with a column \code{Taxa} and the other columns
containing the traits.
By default, the dataset used is the one from Tachet et al. (2010) for macroinvertebrates that
can be retrieved from
\href{https://www.freshwaterecology.info/}{freshwaterecology.info} website
(Schmidt-Kloiber & Hering, 2015).}

\item{group}{Biotic group of interest. Possible values are \code{mi} for macroinvertebrates, \code{mf} for macrophytes and \code{fi} for fish.
The choice will set the right reference database for the specified group.
This option will not be considered if a custom reference database is provided. Default to \code{mi}.}

\item{tax_lev}{Taxonomic level at which the calculation has to be performed.
Default to \code{Taxa}, the maximum taxonomic level is \code{Family}.}

\item{dfref}{Reference dataset as used in the function \code{aggregate_taxa()}.}

\item{filter_by_distance}{Filter the results according to the taxonomic distance. Possible values are \code{pos}, \code{neg} or a positive integer.}

\item{col_blocks}{A vector containing the number of modalities for each trait.}
}
\description{
A function for assigning traits to taxa at different taxonomic levels.
}
\details{
This function allows to obtain missing traits for a target taxon by taking traits from lower or to upper taxomic levels.
For instance, consider the case where the genus \emph{Acroloxus} is present in the user dataset and the species \emph{Acroloxus lacustris}
in the traits dataset. A simple merge would exclude \emph{Acroloxus} from the tha analysis since Acroloxus and \emph{A. lacustris}
would not match. The function assign_traits allows to assign \emph{Acroloxus lacustris} traits to \emph{Acroloxus}.
This function works also in the opposite direction. Consider the case where there are no traits for the target taxon and
the target taxon has not been identified at species level. The function assign_traits will assign the traits of the taxon nearest in the taxonomic tree
(e.g. Tanypodinae traits assign to \emph{Ablabesmyia monilis}). Consider also these examples to understand the behaviour of this
function. For instance \emph{Anabolia lombarda} is present in the user taxomic dataset while only \emph{Anabolia nervosa} and \emph{Anabolia} are present
in the trait database. In this case \code{assign_traits()} will assign only the traits of \emph{Anabolia} to \emph{A. lombarda}.
Moreover, let's assume that \emph{Coelostoma} is present in the user dataset while only \emph{Berosus} and \emph{Crenitis punctatostriata} are present in the traits dataset.
Here \code{assign_traits()} will assign the scores of \emph{Berosus} and \emph{C. punctatostriata} to \emph{Coelostoma} because they belong to the same family and there are no information at family level. \cr \cr
The function \code{assign_traits()} will calculate the taxonomic distance between the target taxa and the taxa used to assign the trait score. This distance
can be positive (e.g. Species to Genus) and negative (e.g. Genus to Species). The distance is measured assigning values as follows:
1 (Species to Genus), 4 (Species to family), -3 (Family to Genus), etc. \code{assign_traits()} considers only the
taxonomic levels from Subspecies to Family (Subspecies, Species, Genus, Tribus, Subfamily, Family).\cr
The function \code{average_traits()} averages traits when traits from multiple taxa are assigned to the same taxon. \cr
Th function \code{sample_traits()} samples traits from one taxon when multiple taxa are assigned to the same taxon.
}
\section{Functions}{
\itemize{
\item \code{average_traits}: average traits values for fuzzy data

\item \code{sample_traits}: sample_traits
}}

\examples{
data(macro_ex)

data_bio <- as_biomonitor(macro_ex)
data_agr <- aggregate_taxa(data_bio)
data_ts <- assign_traits(data_agr)

# select only the nearest traits
data_ts_sub <- manage_traits(data_ts, method = "nearest+-")

# averaging
data_ts_av <- average_traits(data_ts_sub)

# traits random sampling
data_ts_st <- sample_traits(data_ts)
}
\seealso{
\link{aggregate_taxa}
}
