% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmwp.R
\name{bmwp}
\alias{bmwp}
\title{bmwp}
\usage{
bmwp(x, method = "ita", agg = FALSE, exceptions = NULL, traceB = FALSE)
}
\arguments{
\item{x}{Result of \code{aggregate_taxa()}.}

\item{method}{The implementation of BMWP. Possible choices are \code{a} (Armitage et al. 1983), \code{uk} (Davy-Bowker et al. 2010), \code{spa} (MAGRAMA 2011), \code{ita} (Buffagni et al . 2014).
Users can provide their own \code{data.frame} (see examples) with a column called \emph{Taxon} and the column of scores called \emph{Scores}.}

\item{agg}{This option allows the composite family approach. It can be \code{FALSE}, \code{TRUE} or a \code{data.frame}.
If \code{FALSE} no aggregation will be performed, while if \code{TRUE} aggregation will be performed according to the rules described in Details.
A \code{data.frame} containing the aggregation rules can be provided by the user.
This \code{data.frame} needs a column called \emph{Taxon} containing the taxon to aggregate and a column called \emph{Correct_Taxon} with the aggregation specifications.
\code{agg} cannot be \code{TRUE} when a \code{data.frame} is provided as method.}

\item{exceptions}{Taxa that need to be exluded from the calculation.
This option can be useful, for instance, to exclude an alien species belonging to an autochthonous family.}

\item{traceB}{If set to \code{TRUE} a list as specified below will be returned.}
}
\value{
If \code{traceB} is set to \code{TRUE} a list with the following elements will be returned:
\itemize{
\item \code{results} Results of \code{bmwp()}.
\item \code{taxa_df} The \code{data.frame} used for the calculation containing the abundance of taxa receiving a score.
\item \code{composite_taxa} Taxa aggregated following the aggregation rules when \code{agg} is not \code{NULL}.
\item \code{exceptions} A \code{data.frame} containing the containing changes made by excluding the taxa included in \code{exceptions}.
\item \code{parent_child_pairs} For instance in Spanish BMWP both \emph{Ferrissia} and Planorbidae receive a score.
Abundances of the higher taxonomic level need therefore to be adjusted by subtracting the abundances of the lower taxonomic level.
}
}
\description{
This function calculates the \emph{Biological Monitoring Working Party} index following Armitage et al. (1983), Davy-Bowker et al. (2007) and Alba-Tercedor & Sanchez-Ortega (1988) implementations.
}
\details{
BMWP represents the sum of the scores of the taxa receiving a score. Armitage scores are not reliable yet, since taxonomy has to be revised (e.g. Elminthidae are present instead of Elmidae). Davy-Bowker et al. (2010) and Buffagni et al. (2014) implementations take into account composite taxa as follow:
\enumerate{
\item Psychomyiidae (inc. Ecnomidae)
\item Rhyacophilidae (inc. Glossosomatidae)
\item Limnephilidae (inc. Apatanidae)
\item Ancylidae (inc. Acroloxidae)
\item Gammaridae (inc. Crangonyctidae & Niphargidae)
\item Hydrophilidae (inc. Hydraenidae, Helophoridae)
\item Tipulidae (inc. Limoniidae, Pediciidae & Cylindrotomidae)
\item Planariidae (inc. Dugesidae)
\item Hydrobiidae (inc. Bithyniidae)
\item Oligochaeta (all the families)
}

Optional scores provided by the user data.frame needs to be formatted like following:
\tabular{lc}{
Taxon \tab Scores \cr
Aeshnidae \tab 8 \cr
Ancylidae \tab 6 \cr
Aphelocheiridae \tab 10 \cr
Asellidae \tab 3 \cr
Astacidae \tab 8 \cr
}

Optional aggregation \code{data.frame} provided by the user needs to be formatted like following:
\tabular{ll}{
Taxon \tab Correct_Taxon \cr
Glossosomatidae \tab Rhyachopilidae \cr
Apataniidae \tab Limnephilidae \cr
Acroloxidae \tab Ancylidae \cr
Crangonyctidae \tab Gammaridae \cr
Niphargidae \tab Gammaridae \cr
}

The \code{bmwp()} function automatically check for parent-child pairs in the scoring system, see the return the Value section for a definition.
All the information used for \code{bmwp()} calculation can be retrieved with the function \code{show_scores()}.
}
\examples{
data(macro_ex)
data_bio <- as_biomonitor(macro_ex)
data_agr <- aggregate_taxa(data_bio)
bmwp(data_agr)
bmwp(data_agr, method = "spa")
}
\references{
Armitage, P. D., Moss, D., Wright, J. F., & Furse, M. T. (1983). The performance of a new biological water quality score system based on macroinvertebrates over a wide range of unpolluted running-water sites. Water research, 17(3), 333-347.

Davy-Bowker J., Clarke R., Corbin T., Vincent H, Pretty J., Hawczak A., Blackburn J., Murphy J., Jones I., 2008. River Invertebrate Classification Tool. Final report. WFD72C. SNIFFER. 276 pp

MAGRAMA-Ministerio de Agricultura y medio Ambiente (2011) Protocolo de muestreo y laboratorio de fauna bentonica de invertebrados en rios vadeables. ML-Rv-I-2011, Cod, 23 pp.
}
\seealso{
\link{aggregate_taxa} \link{aspt} \link{show_scores}
}
\keyword{aggregate_taxa}
