% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combTaxa.R
\name{combTaxa}
\alias{combTaxa}
\title{combTaxa}
\usage{
combTaxa(x, ntaxa = 2, taxLev = "Family")
}
\arguments{
\item{x}{results of function aggregatoR.}

\item{ntaxa}{number of Taxa to choose.}

\item{taxLev}{taxonimc level on which the calculation has to be made.}
}
\description{
This function returns all the combinations of Taxa at the desired taxonomic resolution.
}
\details{
This function is intended to help the user to identify the best subset of taxa that correlate well with environmental variables.
Metrics based on only certain taxa (EPT, 1-GOLD, etc) are currently used in biomonitoring and their are usually based on autoecological knowledge
of the selected taxa. However, the relationship between an environmental variable and a subset of taxa could exist and not detected due to the actual limitations
about the autoecological knowledge of a group of organsism. With more than 4 taxa the calculations should become infeasible, expecially when the number
of taxa in the user dataset is high. For instance the number of combinations taken 20 at time is nearly 20000.
}
\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
combTaxa(data.agR, taxLev = "Family")
}
\seealso{
\code{\link{aggregatoR}}
}
\keyword{aggregatoR}
