% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combTaxa.R
\name{combTaxa}
\alias{combTaxa}
\title{combTaxa}
\usage{
combTaxa(x, ntaxa = 2, taxLev = "Taxa")
}
\arguments{
\item{x}{result of the function aggregatoR.}

\item{ntaxa}{number of taxa.}

\item{taxLev}{taxonomic level on which the calculation has to be made.}
}
\description{
This function returns all the combinations of n taxa at the desired taxonomic resolution.
}
\details{
This function is intended to help the user to identify the best subset of taxa that correlates with environmental variables.
Metrics based on a specific group of taxa (EPT, 1-GOLD, etc) are currently used in biomonitoring. They rely on available information on the responses
of the selected taxa to target strssors. The relationship between an environmental variable and a subset of taxa could exist and not detected due to the scarce
availaibility of autoecological data. With more than 4 taxa the calculations should become infeasible, expecially when the number
of taxa in the user dataset is high.
}
\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
combTaxa(data.agR, taxLev = "Family")
}
\seealso{
\code{\link{aggregatoR}}
}
\keyword{aggregatoR}
