% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertovegan.R
\name{convertovegan}
\alias{convertovegan}
\title{convertovegan}
\usage{
convertovegan(x, taxLev = "Family")
}
\arguments{
\item{x}{results of aggregatoR function}

\item{taxLev}{taxonomic level of interest. Possible choices are Phylum, Class, Subclass, Order, Family, Subfamily, Tribus, Genus, Species, Subspecies, Taxa}
}
\description{
An utility function to export data to vegan like format
}
\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
data.cv <- convertovegan(data.agR, taxLev = "Family")
}
\seealso{
\code{\link{asBiomonitor}}
}
\keyword{convertovegan}
