% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertobiotic.R, R/convertovegan.R
\name{convertobiotic}
\alias{convertobiotic}
\alias{convertovegan}
\title{convertovegan}
\usage{
convertobiotic(x)

convertovegan(x, taxLev = "Family")
}
\arguments{
\item{x}{result of the aggregatoR function}

\item{taxLev}{taxonomic level of interest. Possible choices are Phylum, Class, Subclass, Order, Family, Subfamily, Tribus, Genus, Species, Subspecies, Taxa}
}
\description{
An utility function to export data to other formats.
}
\details{
\code{convertovegan} converts data to the \code{vegan} package style, with sites on rows and taxa on columns. \code{convertobiotic} converts data to
the format of \code{biotic} package. This function will extract information at the family level, with an exception for Oligochaeta.
In \code{convertobiotic} if both family and above family level taxonomic information of Oligochaeta are concurrently present,
information will be provided at Oligochaeta level.
\code{biomonitoR} automatically identify the taxonomic level at which Oligochaeta are stored (subclass is the default in \code{biomonitoR}).
Be careful, for all the other taxa entered at taxonomic level higher than family will be discarded (e.g. Trombidiformes).
}
\section{Functions}{
\itemize{
\item \code{convertobiotic}: convert data to biotic format
}}

\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
data.cv <- convertovegan(data.agR, taxLev = "Family")
}
\seealso{
\code{\link{asBiomonitor}}
}
\keyword{convertovegan}
