% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csi.R, R/tsi.R
\name{csi}
\alias{csi}
\alias{tsi}
\title{Community trait specialization}
\usage{
csi(x, traitDB = NULL, taxLev = "Taxa", trans = log1p, traceB = FALSE)

tsi(x, traitDB = NULL, taxLev = "Taxa")
}
\arguments{
\item{x}{results of function aggregatoR}

\item{traitDB}{a trait data base with a column \code{Taxa} and the other columns
containing the traits.Please check \link{traitScaling} for building the trait database.}

\item{taxLev}{character string giving the taxonomic level used to retrieve
trait information. Possible levels are \code{"Taxa"}, \code{"Species"}, \code{"Genus"},
\code{"Family"} as returned by the \link{aggregatoR} function.}

\item{trans}{the function used to transform the abundances, by default
\code{log1p}.}

\item{traceB}{when set to TRUE returns a list with the results of the cwm function
and the traits value used for the calculation.}
}
\value{
a table with the CSI values for each trait
}
\description{
\Sexpr[results=rd, stage=render]{ lifecycle::badge("maturing") }

This function calculates the community trait specialization.
}
\details{
This function first takes the abundance table corresponding to the desired
taxonomic level from the \code{x} aggregatoR object.

For each taxon and each trait, a taxon specialization index is calculated
using the following formula:

\deqn{TSI = (sum(c^{2}_tik) - 1/K) / (1 - 1/K)}

with \eqn{c^{2}_tik} being the affinity of taxon \code{t} for the modality \code{k} (among
\code{K}) of trait \code{i}.

Finally, the community trait specialization is calculated for each trait by
averaging the TSIs using the transformed abundances (using the \code{trans}
function) as weigths. Only the taxa having traits contributes to the calculation of
index.
}
\section{Functions}{
\itemize{
\item \code{tsi}: taxon specialization index
}}

\note{
USE WITH CAUTION, STILL IN DEVELOPMENT.
}
\examples{
data(macro_ex)

data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
data.ts <- traitScaling( data.agR )

# averaging
data.ts.av <- traitsMean( data.ts )

# taxon specialization index

tsi(x = data.agR, traitDB = data.ts.av, taxLev = "Taxa")

# community specialization index
csi(x = data.agR, traitDB = data.ts.av, taxLev = "Taxa", trans = log1p)
csi(x = data.agR, traitDB = data.ts.av, taxLev = "Taxa",
    trans = function(x) {
        ifelse(x > 0, 1, 0)
    })
csi(x = data.agR, traitDB = data.ts.av, taxLev = "Genus", trans = log1p)

}
\references{
Mondy, C. P., & Usseglio-Polatera P. (2013) Using Fuzzy-Coded
Traits to Elucidate the Non-Random Role of Anthropogenic Stress in the
Functional Homogenisation of Invertebrate Assemblages. Freshwater Biology,
59 (3), 584-600. https://doi.org/10.1111/fwb.12289.

Tachet, H., Richoux, P., Bournaud, M., & Usseglio-Polatera, P.
(2010). Invertebres d'eau douce: systematique, biologie, ecologie. Paris:
CNRS editions.

Schmidt-Kloiber, A., & Hering, D. (2015). www. freshwaterecology.
info-An online tool that unifies, standardises and codifies more than
20,000 European freshwater organisms and their ecological preferences.
Ecological indicators, 53, 271-282.
}
\seealso{
\link{aggregatoR}
}
