% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csi.R, R/tsi.R
\name{csi}
\alias{csi}
\alias{tsi}
\title{Community trait specialization}
\usage{
csi(x, trait_db = NULL, tax_lev = "Taxa", trans = log1p, traceB = FALSE)

tsi(x, trait_db = NULL, tax_lev = "Taxa")
}
\arguments{
\item{x}{results of function \code{aggregate_taxa()}.}

\item{trait_db}{a trait data base with a column \code{Taxa} and the other columns
containing the traits.Please check \code{assign_traits()} for building the trait database.}

\item{tax_lev}{character string giving the taxonomic level used to retrieve
trait information. Possible levels are \code{"Taxa"}, \code{"Species"}, \code{"Genus"},
\code{"Family"} as returned by \code{aggregate_taxa()}.}

\item{trans}{the function used to transform the abundances, by default
\code{log1p}.}

\item{traceB}{when set to TRUE returns a list with the results of the cwm function
and the traits value used for the calculation.}
}
\value{
a table with the CSI values for each trait
}
\description{
\Sexpr[results=rd, stage=render]{ lifecycle::badge("maturing") }

This function calculates the community trait specialization.
}
\details{
This function first takes the abundance table corresponding to the desired
taxonomic level in \code{x}.

For each taxon and each trait, a taxon specialization index is calculated
using the following formula:

\deqn{TSI = (sum(c^{2}_tik) - 1/K) / (1 - 1/K)}

with \eqn{c^{2}_tik} being the affinity of taxon \code{t} for the modality \code{k} (among
\code{K}) of trait \code{i}.

Finally, the community trait specialization is calculated for each trait by
averaging the TSIs using the transformed abundances (using the \code{trans}
function) as weigths. Only the taxa having traits contributes to the calculation of
index.
}
\section{Functions}{
\itemize{
\item \code{tsi}: taxon specialization index
}}

\examples{
data(macro_ex)

data_bio <- as_biomonitor(macro_ex)
data_agr <- aggregate_taxa(data_bio)
data_ts <- assign_traits(data_agr)

# averaging
data_ts_av <- average_traits(data_ts)

# taxon specialization index

tsi(x = data_agr, trait_db = data_ts_av, tax_lev = "Taxa")

# community specialization index
csi(x = data_agr, trait_db = data_ts_av, tax_lev = "Taxa", trans = log1p)
csi(
  x = data_agr, trait_db = data_ts_av, tax_lev = "Taxa",
  trans = function(x) {
    ifelse(x > 0, 1, 0)
  }
)
csi(x = data_agr, trait_db = data_ts_av, tax_lev = "Genus", trans = log1p)
}
\references{
Mondy, C. P., & Usseglio-Polatera P. (2013) Using Fuzzy-Coded
Traits to Elucidate the Non-Random Role of Anthropogenic Stress in the
Functional Homogenisation of Invertebrate Assemblages. Freshwater Biology,
59 (3), 584-600. https://doi.org/10.1111/fwb.12289.

Tachet, H., Richoux, P., Bournaud, M., & Usseglio-Polatera, P.
(2010). Invertebres d'eau douce: systematique, biologie, ecologie. Paris:
CNRS editions.

Schmidt-Kloiber, A., & Hering, D. (2015). www. freshwaterecology.
info-An online tool that unifies, standardises and codifies more than
20,000 European freshwater organisms and their ecological preferences.
Ecological indicators, 53, 271-282.
}
\seealso{
\link{aggregate_taxa}
}
