% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwm.R
\name{cwm}
\alias{cwm}
\title{Community-Weighted Mean values of traits}
\usage{
cwm(x, trait_db = NULL, tax_lev = "Taxa", trans = log1p, traceB = FALSE)
}
\arguments{
\item{x}{Results of the function \code{aggregate_taxa()}.}

\item{trait_db}{A trait data base with a column \code{Taxa} and the other columns
containing the traits.
By default, the data base used is the one from Tachet \emph{et al} (2010) that
can be retrieved from
\href{https://www.freshwaterecology.info/}{freshwaterecology.info} website
(Schmidt-Kloiber & Hering, 2015).}

\item{tax_lev}{Character string giving the taxonomic level used to retrieve
trait information. Possible levels are \code{"Taxa"}, \code{"Species"}, \code{"Genus"},
\code{"Family"} as returned by the \link{aggregatoR} function.}

\item{trans}{The function used to transform the abundances, by default
\code{log1p}.}

\item{traceB}{When set to TRUE returns a list with the results of the cwm function
and the traits value used for the calculation.}
}
\value{
a table with the CWM values of each trait (trait modality)
}
\description{
This function calculates the community-weighted means of trait categories.
}
\details{
This function first takes the abundance table corresponding to the desired
taxonomic level from the \code{x} aggregatoR object.

Then it searches from the trait data base all the information available at
the desired level and, if required, calculates the corresponding averaged
trait values (e.g. the family trait values are obtained by averaging all the
trait values from taxa with trait information within this family).

Finally, the community mean trait values are calculated using the transformed
abundances (using the \code{trans} function) as weigths
}
\examples{
data(macro_ex)

data_bio <- as_biomonitor(macro_ex)
data_agr <- aggregate_taxa(data_bio)
data_ts <- assign_traits( data_agr )

# averaging
data_ts_av <- average_traits( data_ts )

# community specialization index
cwm(x = data_agr, trait_db = data_ts_av, tax_lev = "Taxa", trans = log1p)
cwm(x = data_agr, trait_db = data_ts_av, tax_lev = "Taxa",
    trans = function(x) {
        ifelse(x > 0, 1, 0)
    })
cwm(x = data_agr, trait_db = data_ts_av, tax_lev = "Genus", trans = log1p)

}
\references{
Tachet, H., Richoux, P., Bournaud, M., & Usseglio-Polatera, P.
(2010). Invertebres d'eau douce: systematique, biologie, ecologie. Paris:
CNRS editions.

Schmidt-Kloiber, A., & Hering, D. (2015). www. freshwaterecology.
info-An online tool that unifies, standardises and codifies more than
20,000 European freshwater organisms and their ecological preferences.
Ecological indicators, 53, 271-282.
}
\seealso{
\link{aggregate_taxa}
}
