% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwm.R
\name{cwm}
\alias{cwm}
\title{Community-Weighted Mean values of traits}
\usage{
cwm(x, traitDB = NULL, taxLev = "Taxa", trans = log1p)
}
\arguments{
\item{x}{results of function aggregatoR}

\item{traitDB}{a trait data base with a column \code{Taxa} and the other columns
containing the traits. If a trait has several modalities they should be
named as follow: TRAIT_MODALITY.

By default (NULL), the data base used is the one from Tachet \emph{et al} (2010)
that can be retrieved from
\href{https://www.freshwaterecology.info/}{freshwaterecology.info} website
(Schmidt-Kloiber & Hering, 2015).}

\item{taxLev}{character string giving the taxonomic level used to retrieve
trait information. Possible levels are \code{"Taxa"}, \code{"Species"}, \code{"Genus"},
\code{"Family"} as returned by the \link{aggregatoR} function.}

\item{trans}{the function used to transform the abundances, by default
\link{log1p}}
}
\value{
a table with the CWM values of each trait (trait modality)
}
\description{
This function calculates the community-weighted means of trait categories.
}
\details{
This function first takes the abundance table corresponding to the desired
taxonomic level from the \code{x} aggregatoR object.

Then it searches from the trait data base all the information available at
the desired level and, if required, calculates the corresponding averaged
trait values (e.g. the family trait values are obtained by averaging all the
trait values from taxa with trait information within this family).

Finally, the community mean trait values are calculated using the transformed
abundances (using the \code{trans} function) as weigths
}
\examples{
data(macro_ex)

data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)

cwm(x = data.agR, taxLev = "Taxa", trans = log1p)
cwm(x = data.agR, taxLev = "Taxa",
    trans = function(x) {
        ifelse(x > 0, 1, 0)
    })
cwm(x = data.agR, taxLev = "Genus", trans = log1p)

}
\references{
Tachet, H., Richoux, P., Bournaud, M., & Usseglio-Polatera, P.
(2010). Invertébrés d'eau douce: systématique, biologie, écologie. Paris:
CNRS editions.

Schmidt-Kloiber, A., & Hering, D. (2015). www. freshwaterecology.
info–An online tool that unifies, standardises and codifies more than
20,000 European freshwater organisms and their ecological preferences.
Ecological indicators, 53, 271-282.
}
\seealso{
\link{aggregatoR}
}
