% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dness.R
\name{dness}
\alias{dness}
\title{dness}
\usage{
dness(
  x,
  complete = FALSE,
  tax_lev = "Genus",
  method = "delta",
  taxa_tree = FALSE
)
}
\arguments{
\item{x}{Result of \code{aggregate_taxa()}.}

\item{complete}{if TRUE unassigned taxon are removed from the taxonomic list of the desired taxonomic level set with tax_lev}

\item{tax_lev}{taxonomic level from which the indices are to be calculated}

\item{method}{available methods are delta (taxonomic diversity), delta.st (taxonomic distinctness) and delta.bin (variation in taxonomic distinctness)}

\item{taxa_tree}{if TRUE dness return a list where the first element is the taxonomic diversity, taxonomic distinctness or variation of taxonomic distinctness and the second elemnt the taxonomic tree}
}
\description{
\Sexpr[results=rd, stage=render]{ lifecycle::badge("experimental") }

This function calculates the taxonomic diversity, taxonomic distinctness and variation of taxonomic distinctness.
}
\details{
see Clarke and Warwick (1998) an Clarke and Warwick (2001).
}
\examples{
data(oglio)
data_bio <- as_biomonitor(oglio, group = "mf")
data_agr <- aggregate_taxa(data_bio)
dness(data_agr, complete = TRUE)
}
\references{
Clarke, K. R., & Warwick, R. M. (1998). A taxonomic distinctness index and its statistical properties. Journal of applied ecology, 35(4), 523-531.

Clarke, K. R., & Warwick, R. M. (2001). A further biodiversity index applicable to species lists: variation in taxonomic distinctness. Marine ecology Progress series, 216, 265-278.
}
\seealso{
\code{\link{aggregatoR}}
}
\keyword{ept}
