% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epsi.R
\name{epsi}
\alias{epsi}
\title{e-psi}
\usage{
epsi(x, taxLev = "Family", abucl = c(0, 9, 99, 999), composite = F)
}
\arguments{
\item{x}{results of aggregatoR function}

\item{taxLev}{currently only the option "Family" is allowed}

\item{abucl}{Log abundance categories. Treshold are set to 0, 9, 99, 999.}
}
\description{
This function calculates Empyrically-weighted Proportion of Sediment-sensitive Invertebrates index (ePSI) according to the most recent version used in UK.
}
\details{
ePSI implementation take into account composite taxa as follow:
\enumerate{
\item Tipulidae (inc. Limoniidae, Pediciidae & Cylindrotomidae)
\item Siphlonuridae (inc. Ameletidae)
\item Hydrophilidae (inc. Georissidae, Helophoridae & Hydrochidae)
}
Ancylus (0.51) has a different score compared to the other Planorbidae (0). Take it into account when interpreting the results.
Scores used for epsi calculation can be explored with the function code{\link{showscores}}.
}
\section{Acknowledgements}{
 We thank Carol Fitzpatrick, Richard Chadd, Judy England and Rachel Stubbington for providing us with the most updated ePSI scores and algorithms.
}

\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
data.psi <- epsi(data.agR, taxLev = "Family", composite = F)
}
\references{
Turley MD, Bilotta GS, Chadd RP, Extence CA, Brazier RE, Burnside NG, Pickwell AG. 2016. A sediment-specific family-level biomonitoring tool to identify the impacts of fine sediment in temperate rivers and streams. Ecological Indicators 70, 151-165.

Turley MD, Bilotta GS, Krueger T, Brazier RE, Extence CA. 2015. Developing an improved biomonitoring tool for fine sediment: combining expert knowledge and empirical data. Ecological indicators 54, 82-86.
}
\seealso{
\code{\link{asBiomonitor}}
}
\keyword{epsi}
