% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ept.R
\name{ept}
\alias{ept}
\title{ept}
\usage{
ept(x, tax_lev = "Taxa")
}
\arguments{
\item{x}{Result of the function \code{aggregate_taxa}.}

\item{tax_lev}{The taxonomic level for calculating EPT richness.}
}
\description{
This function calculates the richness of Ephemeroptera, Plecotera and Trichoptera (EPT) taxa at the desired taxonomic level.
}
\details{
\code{tax_lev} must be finer than Order (e.g. Species, Genus, Family ).
The \code{ept()} function works if Ephemeroptera, Trichoptera and Plecoptera are set as orders in the reference database. Otherwise try with \code{\link{ricTax}}.
}
\examples{
data(macro_ex)
data_bio <- as_biomonitor(macro_ex)
data_agr <- aggregate_taxa(data_bio)
ept(data_agr)

# ept should return the same results as the function get_taxa_richness
ept( data_agr , tax_lev = "Family" )
get_taxa_richness( data_agR , taxa = c( "Ephemeroptera" , "Plecoptera" , "Trichoptera" ) , tax_lev = "Family" )
}
\seealso{
\code{\link{aggregatoR}} \code{\link{ricTax}}
}
\keyword{ept}
