% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ept.R
\name{ept}
\alias{ept}
\title{ept}
\usage{
ept(x, taxLev = "Family")
}
\arguments{
\item{x}{results of function aggregatoR}
}
\description{
This function calculates EPT richnes
}
\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
ept(data.bio)
}
\seealso{
\code{\link{aggregatoR}}
}
\keyword{ept}
