% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eptd.R
\name{eptd}
\alias{eptd}
\title{eptd}
\usage{
eptd(x)
}
\arguments{
\item{x}{results of function aggregatoR}
}
\description{
This function calculates the log10(Sel_EPTD + 1) metric.
}
\details{
log10(Sel_EPTD + 1) the base-10 logarithm of the abundance of the selected EPTD families plus 1.
}
\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
eptd(data.agR)
}
\seealso{
\code{\link{aggregatoR}}
}
\keyword{ept}
