% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_disp.R
\name{f_disp}
\alias{f_disp}
\title{Functional dispersion}
\usage{
f_disp(
  x,
  trait_db = NULL,
  tax_lev = "Taxa",
  type = NULL,
  traitSel = FALSE,
  col_blocks = NULL,
  nbdim = 2,
  distance = "gower",
  zerodist_rm = FALSE,
  correction = "none",
  traceB = FALSE,
  set_param = list(max_nbdim = 15, prec = "Qt", tol = 1e-07, cor.zero = TRUE)
)
}
\arguments{
\item{x}{results of function \code{aggregate_taxa()}.}

\item{trait_db}{a trait database. Can be a \code{data.frame} ot a \code{dist} object.
Taxonomic level of the traits database must match those of the taxonomic database.
No automatic check is done by the \code{function}.}

\item{tax_lev}{character string giving the taxonomic level used to retrieve
trait information. Possible levels are \code{"Taxa"}, \code{"Species"}, \code{"Genus"},
\code{"Family"} as returned by the \link{aggregatoR} function.}

\item{type}{the type of variables speciefied in \code{trait_db}.
Must be one of \code{F}, fuzzy, or \code{C}, continuous.
If more control is needed please consider to provide \code{trait_db} as a \code{dist} object.
It works only when \code{trait_db} is a \code{data.frame}, otherwise ingored.}

\item{traitSel}{interactively select traits.}

\item{col_blocks}{A vector that contains the number of modalities for each trait.
Not needed when \code{euclidean} distance is used.}

\item{nbdim}{number of dimensions for the multidimensional functional spaces.
We suggest to keep \code{nbdim} as low as possible.
By default \code{biomonitoR} set the number of dimensions to 2. Select \code{auto} if you want the automated selection
approach according to Maire et al. (2015).}

\item{distance}{to be used to compute functional distances, \code{euclidean} or \code{gower}. Default to \code{gower}.}

\item{zerodist_rm}{If \code{TRUE} aggregates taxa with the same traits.}

\item{correction}{Correction methods for negative eigenvalues, can be one of \code{none}, \code{lingoes}, \code{cailliez}, \code{sqrt} and \code{quasi}.
Ignored when type is set to \code{C}.}

\item{traceB}{if \code{TRUE} ffrich will return a list as specified in details.}

\item{set_param}{a list of parameters for fine tuning the calculations.
\code{max_nbdim} set the maximum number of dimension for evaluating the quality of the functional space.
\code{prec} can be \code{Qt} or \code{QJ}, please refere to the \code{convhulln} documentation for more information.
Deafault to \code{QJ}, less accurate but less prone to errors.
\code{tol} a tolerance threshold for zero, see the function \code{is.euclid}, \code{lingoes} and \code{cailliez} from the \code{ade4} for more details. Default to 1e-07.
\code{cor.zero} = \code{TRUE} if TRUE, zero distances are not modified. see the function \code{is.euclid}, \code{lingoes} and \code{cailliez} from the \code{ade4} for more details. Default to \code{TRUE}.}
}
\value{
a vector with fuzzy functional richness results.
\enumerate{
\item \strong{results}: results of \code{f_disp()};
\item \strong{traits}: a data.frame containing the traits used for the calculations;
\item \strong{taxa}: a data.frame conaining the taxa used for th calculations;
\item \strong{nbdim}: number of dimensions used after calculatin the quality of functional spaces according to Maire et al., (2015);
\item \strong{correction}: the type of correction used.
\item \strong{NA_detection}: a data.frame containing taxa on the first column and the corresponding trais with NAs on the second column.
\item \strong{duplicated_traits}: if present, list the taxa with the same traits.
}
}
\description{
\Sexpr[results=rd, stage=render]{ lifecycle::badge("maturing") }

This function calculates the functional dispersion.
}
\examples{
data(macro_ex)

data_bio <- as_biomonitor(macro_ex)
data_agr <- aggregate_taxa(data_bio)
data_ts <- assign_traits( data_agr )
# averaging
data_ts_av <- average_traits( data_ts )

col_blocks <- c( 8, 7, 3, 9, 4, 3, 6, 2, 5, 3, 9, 8, 8, 5, 7, 5, 4, 4, 2, 3, 8 )

f_disp( data_agr , trait_db = data_ts_av , type = "F" , col_blocks = col_blocks )
f_disp( data_agr , trait_db = data_ts_av , type = "F" , col_blocks = col_blocks ,
       nbdim = 10 , correction = "cailliez" )

library( ade4 )

rownames( data_ts_av ) <- data_ts_av$Taxa
traits_prep <- prep.fuzzy( data_ts_av[ , -1 ], col.blocks = col_blocks )

traits_dist <- ktab.list.df( list( traits_prep ) )
traits_dist <- dist.ktab( traits_dist , type = "F" )

f_disp( data_agr , trait_db = traits_dist   )


}
\references{
Barber, C. B., Dobkin, D. P., & Huhdanpaa, H. (1996).
The quickhull algorithm for convex hulls. ACM Transactions on
Mathematical Software (TOMS), 22(4), 469-483.

Cornwell, W. K., Schwilk, D. W., & Ackerly, D. D. (2006).
A trait-based test for habitat filtering: convex hull volume.
Ecology, 87(6), 1465-1471

Maire, E., Grenouillet, G., Brosse, S., & Villeger, S. (2015).
How many dimensions are needed to accurately assess functional diversity?
A pragmatic approach for assessing the quality of functional spaces. Global
Ecology and Biogeography, 24(6), 728-740.

Mason, N. W., Mouillot, D., Lee, W. G., and
Wilson, J. B. (2005). Functional richness, functional evenness and functional
divergence: the primary components of functional diversity. Oikos, 111(1),
112-118.

Villeger, S., Mason, N. W., & Mouillot, D.
(2008). New multidimensional functional diversity indices for a
multifaceted framework in functional ecology. Ecology, 89(8), 2290-2301.
}
\seealso{
\link{aggregate_taxa}
}
