% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffdisp.R
\name{ffdisp}
\alias{ffdisp}
\title{Fuzzy coded functional dispersion}
\usage{
ffdisp(x, traitDB = NULL, agg = FALSE, dfref = NULL,
  traitSel = FALSE, colB = NULL, taxLev = "Taxa", traceB = FALSE)
}
\arguments{
\item{x}{results of function aggregatoR}

\item{traitDB}{a trait data base with a column \code{Taxa} and the other columns
containing the traits. Needed only if users use their own trait database.  If a trait has several modalities they should be
named as follow: TRAIT_MODALITY.

By default, the data base used is the one from Tachet \emph{et al} (2010) that
can be retrieved from
\href{https://www.freshwaterecology.info/}{freshwaterecology.info} website
(Schmidt-Kloiber & Hering, 2015).}

\item{agg}{should ffrich aggregate user's traitDB of higher taxonomic level? TRUE to aggregate, otherwise FALSE.
For instance, if user's traitDB has both Halesus and Limnephilidae, ffred will aggregate traits value if agg = TRUE.}

\item{dfref}{reference database to be used when a custom trait database is provided and agg = TRUE.
It should be the same reference database used in \link{asBiomonitor} when dfref = TRUE.}

\item{traitSel}{interactively select traits. See details for more info.}

\item{colB}{A vector that contains the number of modalities for each trait. Needed only if users use their own trait database.}

\item{taxLev}{character string giving the taxonomic level used to retrieve
trait information. Possible levels are \code{"Taxa"}, \code{"Species"}, \code{"Genus"},
\code{"Family"} as returned by the \link{aggregatoR} function.}

\item{traceB}{if TRUE ffred will return a list with 2 elements, the first being the ffrich values and the second the database used for the calculation. Useful to check missing taxa.}
}
\value{
A vector with functional dispersion results.\cr
If traceB is set to TRUE a list is provided with:
\enumerate{
\item \strong{results}: results of the ffred function;
\item \strong{traits}: a data.frame containing the traits used for the calculations;
\item \strong{taxa}: a data.frame conaining the taxa used for th calculations;
\item \strong{taxa_not_used}: a vector conaining the names of the taxa exluded from the calculations. Taxa are excluded from the calculations if they have NA values at least in one of the trait modalities;
\item \strong{problematic traits}: traits containing NAs for \emph{taxa_not_used}
\item \strong{traits_not_used}: trait modalities excluded from the calculations because they have 0 value for each species.
}
}
\description{
FDis is defined as the mean distance in multidimensional trait space of individual species to the centroid of all species (Laliberte & Legendre, 2010).
}
\details{
Taxa without traits assigned in the trait database are removed from both the trait and abundance databases.
}
\note{
USE WITH CAUTION, STILL IN DEVELOPMENT.
}
\examples{
data(macro_ex)

data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
ffdisp(data.agR)

}
\references{
Laliberte, E., & Legendre, P. (2010). A distance-based framework for measuring functional diversity from multiple traits. Ecology, 91(1), 299-305.
}
\seealso{
\link{aggregatoR}
}
