% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffrich.R
\name{ffrich}
\alias{ffrich}
\title{Fuzzy coded functional richness}
\usage{
ffrich(x, traitDB = NULL, agg = FALSE, traitSel = FALSE,
  taxLev = "Family", traceB = FALSE, nbdim = 7, metric = "Gower",
  corr_method = FALSE)
}
\arguments{
\item{x}{results of function aggregatoR}

\item{traitDB}{a trait data base with a column \code{Taxa} and the other columns
containing the traits. If a trait has several modalities they should be
named as follow: TRAIT_MODALITY.

By default, the data base used is the one from Tachet \emph{et al} (2010) that
can be retrieved from
\href{https://www.freshwaterecology.info/}{freshwaterecology.info} website
(Schmidt-Kloiber & Hering, 2015).}

\item{agg}{should ffrich aggregate user's traitDB of higher taxonomic level? TRUE to aggregate, otherwise FALSE.
For instance, if user's traitDB has both Halesus and Limnephilidae, ffrich will aggregate traits value if ADD = TRUE.}

\item{traitSel}{interactively select traits.}

\item{taxLev}{character string giving the taxonomic level used to retrieve
trait information. Possible levels are \code{"Taxa"}, \code{"Species"}, \code{"Genus"},
\code{"Family"} as returned by the \link{aggregatoR} function.}

\item{traceB}{if TRUE ffrich will return a list with 2 elements, the first being the ffrich values and the second the database used for the calculation. Useful to check missing taxa.}

\item{nbdim}{maximum number of dimensions for multidimensional functional spaces. By default, nbdim=7                   ##
Final number of dimensions depends on the number of positive eigenvalues (after correction) obtained with PCoA}

\item{metric}{to be used to compute functional distance, "Euclidean" or "Gower" (=default)}

\item{corr_method}{TRUE will aplly the Cailliez correction.}
}
\value{
a vector with fuzzy functional richness results.
}
\description{
This function calculates the functional richness based on trait category.
}
\details{
This function first takes the abundance table corresponding to the desired
taxonomic level from the \code{x} aggregatoR object.

Then it searches from the trait data base all the information available at
the desired level and, if required, calculates the corresponding averaged
trait values (e.g. the family trait values are obtained by averaging all the
trait values from taxa with trait information within this family).
}
\note{
USE WITH CAUTION, STILL IN DEVELOPMENT.
}
\examples{
data(macro_ex)

data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)

cwm(x = data.agR, taxLev = "Taxa")

}
\references{
Tachet, H., Richoux, P., Bournaud, M., & Usseglio-Polatera, P.
(2010). Invertebres d'eau douce: systematique, biologie, ecologie. Paris:
CNRS editions.

Schmidt-Kloiber, A., & Hering, D. (2015). www. freshwaterecology.
info-An online tool that unifies, standardises and codifies more than
20,000 European freshwater organisms and their ecological preferences.
Ecological indicators, 53, 271-282.
}
\seealso{
\link{aggregatoR}
}
