% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_info.R
\name{general_info}
\alias{general_info}
\title{general_info}
\usage{
general_info(x, taxalist = FALSE)
}
\arguments{
\item{x}{results of the function \code{aggregate_taxa()}.}

\item{taxalist}{if \code{TRUE} returns the list of taxa for each taxonomic level.}
}
\description{
Function to calculate the overall richness at different taxonomic level.
}
\examples{
data(macro_ex)
data_bio <- as_biomonitor(macro_ex)
data_agr <- aggregate_taxa(data_bio)
general_info(data_agr)
}
\seealso{
\code{\link{as_biomonitor}}
}
\keyword{general_info}
