% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gbif_taxa.R
\name{get_gbif_taxa_tree}
\alias{get_gbif_taxa_tree}
\title{Create \code{biomonitoR} dataset from GBIF}
\usage{
get_gbif_taxa_tree(x, ref_from_tree = FALSE)
}
\arguments{
\item{x}{Vector containing your taxa list.}

\item{ref_from_tree}{Create a reference database in the \code{biomonitoR} format. See \link{ref_from_tree}.}
}
\description{
Function to create a reference dataset using API from the Global Biodiversity Information Facility (GBIF).
}
\examples{
\dontrun{

data(macro_ex)
dfref_gbif <- get_gbif_taxa_tree(macro_ex[, "Taxa"])
data_asb <- as_biomonitor(macro_ex, dfref = dfref_gbif$taxonomy)
data_agg <- aggregate_taxa(data_asb)
}
}
\seealso{
\link{as_biomonitor}
}
\keyword{as_biomonitor}
