% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxa_abundance.R
\name{get_taxa_abundance}
\alias{get_taxa_abundance}
\title{get_taxa_abundance}
\usage{
get_taxa_abundance(x, taxa = NULL, rel = FALSE)
}
\arguments{
\item{x}{Result of \code{aggregate_taxa()}.}

\item{taxa}{A taxon or a vector of taxa.}

\item{rel}{If \code{TRUE} calculates relative abundance. Default to \code{FALSE}.}
}
\description{
Calculates the absolute or relative abundance of a taxon or of a set taxa.
}
\details{
This function does not check for parent-child pairs. For instance if the vector of taxa contains both the genus \emph{Baetis} and the
family Baetidae, the abundance of the genus \emph{Baetis} will be double counted. Check the function \code{solve_ambiguous()} for a solution
to this problem.
}
\examples{
data(macro_ex)
data_bio <- as_biomonitor(macro_ex)
data_agr <- aggregate_taxa(data_bio)
get_taxa_abundance(data_agr, taxa = "Ephemeroptera")
get_taxa_abundance(data_agr, taxa = c("Setodes", "Orthocladiinae"), rel = TRUE)
}
\seealso{
\link{aggregate_taxa} \link{solve_ambiguous} \link{get_taxa_richness}
}
\keyword{aggregatoR}
