% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold.R
\name{gold}
\alias{gold}
\title{gold}
\usage{
gold(x)
}
\arguments{
\item{x}{results of function aggregatoR}
}
\description{
This function calculates the 1 - GOLD metric, where GOLD stands for Gastropoda, Oligochaeta and Diptera.
}
\details{
The metric 1 - GOLD is calculated as 1 minus the relative abundance of Gastropoda, Oligochaeta and Diptera. If a custom database is provided (see \code{\link{aggregatoR}}) please be sure that Gastropoda and Oligochaeta are submitted as Class and Diptera as Order, otherwise the gold calculation will be meaningless.
}
\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
gold(data.agR)
}
\seealso{
\code{\link{aggregatoR}}
}
\keyword{ept}
