% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibmr.R
\name{ibmr}
\alias{ibmr}
\title{Biologique Macrofitique en Riviere index}
\usage{
ibmr(
  x,
  method = "ita",
  exceptions = NULL,
  coverage_coef = FALSE,
  coverage_classes = c(0.1, 1, 10, 50)
)
}
\arguments{
\item{x}{Result of \code{aggregate_taxa()}.}

\item{method}{The implementation of IBMR. the default method is \code{ita} (Minciardi et al., 2009).
Users can provide their own data.frame (see examples) with a column called \emph{Taxon} and the column of scores called \emph{Scores}.}

\item{exceptions}{Taxa that need to be excluded from the calculation.
This option can be useful, for instance, to exclude an alien species belonging to an autochthonous family.}

\item{coverage_coef}{Are data already represented with coverage coefficients? Set to \code{TRUE} if this is not the case.}

\item{coverage_classes}{Thresholds to transform abundances into coverage classes.}
}
\description{
\Sexpr[results=rd, stage=render]{ lifecycle::badge("maturing") }

This function calculates the IBMR index following Minciardi et al. (2009).
}
\details{
The \code{traceB} argument will be implemented soon.
}
\examples{
data(oglio)
oglio_asb <- as_biomonitor(oglio, group = "mf", FUN = bin)
oglio_agg <- aggregate_taxa(oglio_asb)
ibmr(oglio_agg)
}
\references{
Minciardi, M. R., Spada, C. D., Rossi, G. L., Angius, R., Orru, G., Mancini, L., Pace, G., Marcheggiani, S., Puccinelli, C. (2009). Metodo per la valutazione e la classificazione dei corsi d'acqua utilizzando la comunita delle macrofite acquatiche. ENEA Rapporto Tecnico RT/2009/23/ENEA.
}
\seealso{
\link{aggregate_taxa}
}
\keyword{aspt}
