% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.R
\name{life}
\alias{life}
\title{life}
\usage{
life(x, method = "Family")
}
\arguments{
\item{x}{results of function aggregatoR function}

\item{method}{possible choices}
}
\description{
This function calculates LIFE index according to Extence et al. (1999).
}
\details{
life currently calculates life index to family and species level, following the nomenclature used by Extence et al. (1999). Species level nomenclature of Extence et al. (1999) is outdated, an updated version will be released soon in biomonitoR.
}
\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
data.life <- life(data.agR, method = "Family")
}
\references{
Extence CA, Balbi DM, Chadd RP. 1999. River flow indexing using British benthic macroinvertebrates: a framework for setting hydroecological objectives. Regulated Rivers: Research and Management 15: 543–574.
}
\seealso{
\code{\link{asBiomonitor}}
}
\keyword{life}
