% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.R
\name{life}
\alias{life}
\title{life}
\usage{
life(x, taxLev = "Family", version = "extence", composite = FALSE,
  abucl = c(0, 9, 99, 999, 9999))
}
\arguments{
\item{x}{results of function aggregatoR function}

\item{taxLev}{currently only the option "Family" is enabled}

\item{version}{possible choices are "extence" and "life_2017"}

\item{composite}{if TRUE composite families as listed in the details section are used}

\item{abucl}{abundance threshold. Default 0, 9, 99, 999, 9999}
}
\description{
This function calculates LIFE index according to most recent version used in UK.
}
\details{
Lotic-invertebrate Index for Flow (LIFE) was originally proposed by Extence et al. (1999). biomonitoR implements the Extence et al. (1999) version called "extence" and the version currently used in UK called "life_2017".
If composite is set to T the following composite families are used for extence

\enumerate{
\item Psychomyiidae (inc. Ecnomidae)
\item Rhyacophilidae (inc. Glossomatidae)
\item Ancylidae (inc. Acroloxidae)
\item Gammaridae (inc. Crangonyctidae)
\item Planariidae (inc. Dugesidae)
\item Hydrobiidae (inc. Bithyniidae)
}

while for "life_2017" the following are used:

\enumerate{
\item Hydrophilidae (inc. Georissidae, Hlophoridae, Hydrochidae)

}
Scores used for life calculation can be explored with the function code{\link{showscores}}.
}
\section{Acknowledgements}{
 We thank Carol Fitzpatrick, Richard Chadd, Judy England and Rachel Stubbington for providing us with the most updated LIFE scores and algorithms.
}

\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
data.life <- life(data.agR, taxLev = "Family", composite = FALSE)
}
\references{
Extence CA, Balbi DM, Chadd RP. 1999. River flow indexing using British benthic macroinvertebrates: a framework for setting hydroecological objectives. Regulated Rivers: Research and Management 15: 543-574.
}
\seealso{
\code{\link{asBiomonitor}}
}
\keyword{life}
