% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_ex.R
\docType{data}
\name{macro_ex}
\alias{macro_ex}
\title{Ephemeroptera toy example}
\format{A data frame with 4 rows and 3 variables:
\describe{
  \item{Taxa}{name of ephemeroptera taxa}
  \item{ST1}{abundance in site 1}
  \item{ST2}{abundance in site 2}
}}
\usage{
macro_ex
}
\description{
A toy axample to illustrate the biomonitoR package use
diamonds.
}
\keyword{datasets}
