% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_asb.R
\name{plot.asb}
\alias{plot.asb}
\title{Plot objects generated with as_biomonitor}
\usage{
\method{plot}{asb}(
  x,
  parent = "Order",
  child = "Family",
  type = "pa",
  remove_empty_child = FALSE,
  trans = NULL,
  ...
)
}
\arguments{
\item{x}{Result of \code{as_biomonitor()}.}

\item{parent}{Name of the coarser taxonomic level.}

\item{child}{Name of the finer taxonomic level.}

\item{type}{Plot present-absence data with \code{pa}, abundance data with \code{abundance} and
detection probabilities with \code{frequency}.}

\item{remove_empty_child}{Remove empty level of the child taxonomic level.}

\item{trans}{Transformation for abundance data.}

\item{...}{Further arguments to be passed to \code{plot}.}
}
\description{
Plot 2 taxonomic levels using a sunburst visualization.
}
\examples{
data(macro_ex)
data_bio <- as_biomonitor(macro_ex)
plot(data_bio)
}
