% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickRename.R
\name{quickRename}
\alias{quickRename}
\title{quickRename}
\usage{
quickRename(x, group = "mi", write.table = FALSE)
}
\arguments{
\item{x}{a data.frame as specified in details}

\item{group}{biotic group of interest. Possible values are "mi" for macroinvertebrates and "mf" for macrophytes.}

\item{write.table}{if T quickRename will save a csv file with changes provided by the user}
}
\description{
This function allow the user to change the wrong taxa names.
}
\details{
data.frame must have a column called "Taxa" where put species, genus or family names. See data(macro_ex) for an example dataset.\cr
The function quickRename will suggest correct name and allow the user to insert a name (Enter taxon name).
}
\examples{
data(macro_ex)
macro_ex.mod <- quickRename(macro_ex, group = "mi")
}
\seealso{
\code{\link{asBiomonitor}}
}
\keyword{asBiomonitor}
