% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_from_tree.R
\name{ref_from_tree}
\alias{ref_from_tree}
\title{Reference from taxonomic tree}
\usage{
ref_from_tree(x, group = "none")
}
\arguments{
\item{x}{Taxonomic tree. See \code{\link{Tree}} for an example.}

\item{group}{Merge the user dataset with one of the \code{biomonitoR} reference databases, default to \code{none}.
If duplicated Taxa names are present this function keeps the name provided by the user.}
}
\description{
This function transforms a taxonomic tree to a reference database suitable for \code{biomonitoR}.
}
\examples{
data(Tree)
ref_custom <- ref_from_tree(Tree)
}
\keyword{ref_from_tree}
