% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refFromTree.R
\name{refFromTree}
\alias{refFromTree}
\title{refFromTree}
\usage{
refFromTree(x, group = "none")
}
\arguments{
\item{x}{taxonomic tree. See \code{\link{Tree}} for an example.}

\item{group}{merge the user database with one of the biomonitoR reference databases, default to \code{none}.
If duplicated Taxa names are present this function keeps the name provided by the user.
Check the reliability of results when using group = \code{mi} for macroinvertebrates or group = \code{mf} for macrophytes.}
}
\description{
\Sexpr[results=rd, stage=render]{ lifecycle::badge("deprecated") }

This function transforms a taxonomic tree to a reference database suitable for biomonitoR.
}
\keyword{refFromTree}
