% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refFromTree.R
\name{refFromTree}
\alias{refFromTree}
\title{refFromTree}
\usage{
refFromTree(x)
}
\arguments{
\item{x}{taxonomic tree}
}
\description{
This function transforms a taxonomic tree to a reference database suitable for biomonitoR.
}
\examples{
data(Tree)
ref_custom <- refFromTree(Tree)
}
\keyword{refFromTree}
