% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refFromTree.R
\name{refFromTree}
\alias{refFromTree}
\title{refFromTree}
\usage{
refFromTree(x, group = "none")
}
\arguments{
\item{x}{taxonomic tree}

\item{group}{merge the user database with biomonitoR reference databases, default to "none". If duplicated Taxa names are present this function keeps the name provided by the user. Check the reliability of results when usin group = "mi" or group = "mf"}
}
\description{
This function transforms a taxonomic tree to a reference database suitable for biomonitoR.
}
\examples{
data(Tree)
ref_custom <- refFromTree(Tree)
}
\keyword{refFromTree}
