% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_taxa.R
\name{remove_taxa}
\alias{remove_taxa}
\title{remove_taxa}
\usage{
remove_taxa(x, taxa = NULL)
}
\arguments{
\item{x}{Result of \code{aggregate_taxa()}.}

\item{taxa}{A taxon or a vector of taxa.}
}
\description{
\Sexpr[results=rd, stage=render]{ lifecycle::badge("experimental") }

Remove a taxon or a set of taxa and create a new \code{data.frame}.
}
\details{
This function does not check for parent-child pairs but it should handle them pretty well.
}
\examples{
data(macro_ex)
data_bio <- as_biomonitor(macro_ex)
data_agr <- aggregate_taxa(data_bio)
remove_taxa(data_agr, taxa = "Ephemeroptera")
}
\seealso{
\code{\link{aggregatoR}} \code{\link{ambiguousSolver}}
}
\keyword{aggregate_taxa}
