% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ricTax.R
\name{ricTax}
\alias{ricTax}
\title{ricTax}
\usage{
ricTax(x, taxa = NULL, taxLev = NULL)
}
\arguments{
\item{x}{results of function aggregatoR.}

\item{taxa}{a Taxon or a vector of taxa.}

\item{taxLev}{taxonomic level at which the richness has to be calculated.}
}
\description{
This function calculates the absolute richness of a Taxon or of a set Taxa at a user provided taxonomic level.
}
\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
ricTax(data.agR, taxa = "Ephemeroptera", taxLev = "Family")
ricTax(data.agR, taxa = c("Plecopotera", "Ephemeroptera"), taxLev = "Family")
# The following line give the same result
ricTax(data.agR, taxa = c("Plecopotera", "Ephemeroptera"), taxLev = c("Family", "Family"))
}
\seealso{
\code{\link{aggregatoR}}
}
\keyword{aggregatoR}
