% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ricTax.R
\name{ricTax}
\alias{ricTax}
\title{ricTax}
\usage{
ricTax(x, taxa = NULL, taxLev = NULL)
}
\arguments{
\item{x}{result of the function aggregatoR.}

\item{taxa}{a taxon or a vector of taxa.}

\item{taxLev}{taxonomic level at which the richness has to be calculated. It could be also a vector of taxonomic levels.}
}
\description{
This function calculates the absolute richness of a taxon or of a set taxa at the taxonomic level provided by the user.
}
\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
ricTax(data.agR, taxa = "Ephemeroptera", taxLev = "Family")
}
\seealso{
\code{\link{aggregatoR}}
}
\keyword{aggregatoR}
