% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon.R
\name{shannon}
\alias{shannon}
\title{shannon}
\usage{
shannon(x, base = 2, taxLev = "Family")
}
\arguments{
\item{x}{results of function aggregatoR}

\item{base}{the base of the logarithm}

\item{taxaLev}{taxonimc level on which the calculation has to be made.}
}
\description{
Functions for calculating shannon, simpson, margalef and menhinick indexes.
}
\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
shannon(data.agR)
}
\seealso{
\code{\link{aggregatoR}}
}
\keyword{margalef,}
\keyword{menhinick}
\keyword{shannon,}
\keyword{simpson,}
