% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showscores.R
\name{showscores}
\alias{showscores}
\title{showscores}
\usage{
showscores(x, writecsv = F)
}
\arguments{
\item{x}{name of the scores to print. Allowed names are "aspt_b", "aspt_i", "aspt_h", "life_scores_fam", "whpt_scores_fam", "psi_scores_fam" and "psi_scores_fam"}

\item{writecsv}{if TRUE the scores are saved in the working directory.}
}
\description{
This function print the scores used for the calculation of several indices in biomonitoR.
}
\examples{
showscores("aspt_i", writecsv = F)
}
\seealso{
\code{\link{aspt}}, \code{\link{bmwp}}, \code{\link{life}}, \code{\link{whpt}}
}
\keyword{asBiomonitor}
