% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showscores.R
\name{showscores}
\alias{showscores}
\title{showscores}
\usage{
showscores(index = NULL, method = NULL)
}
\arguments{
\item{index}{name of the index for which information are needed.}

\item{method}{method of the index specified in \code{index}.}
}
\description{
\Sexpr[results=rd, stage=render]{ lifecycle::badge("deprecated") }

This function print the databases used for the calculation of several indices implemented in biomonitoR.
}
\seealso{
\code{\link{aspt}}, \code{\link{bmwp}}, \code{\link{life}}, \code{\link{whpt}}, \code{\link{psi}}, \code{\link{epsi}}
}
\keyword{asBiomonitor}
