% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speNumb.R
\name{speNumb}
\alias{speNumb}
\alias{genNumb}
\alias{famNumb}
\alias{taxNumb}
\alias{abu}
\title{speNumb}
\usage{
speNumb(x)
}
\arguments{
\item{x}{results of function aggregatoR}
}
\description{
Functions for calculating species, genus, family richness and abundance
}
\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
genNumb(data.agR)
}
\seealso{
\code{\link{aggregatoR}}
}
\keyword{abu}
\keyword{famNumb,}
\keyword{genNumb,}
\keyword{ordNumb,}
\keyword{speNumb,}
\keyword{taxNumb,}
