% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_frame_asb.R
\name{subset.asb}
\alias{subset.asb}
\title{subset.asb}
\usage{
\method{subset}{asb}(x, ..., taxa, exclude = NULL)
}
\arguments{
\item{x}{results of function \code{asBiomonitor}.}

\item{...}{further arguments to be passed to or from other methods.}

\item{taxa}{a taxon or a vector of taxa to retain.}

\item{exclude}{a taxon or a vector of taxa to exclude. For example, this option is useful to exclude alien taxa.}
}
\description{
\code{summary} method for class \code{asb}.
}
\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)

# select EPT Taxa (Ephemeroptera, Plecoptera and Trichoptera)

subset(data.bio, taxa = c( "Ephemeroptera" , "Plecoptera" , "Trichoptera" ) )

# select Trichoptera excluding the trichopteran family Hydropsychidae

tricho <- subset(data.bio, taxa = "Trichoptera" , exclude = "Hydropsychidae" )
tricho.agg <- aggregatoR( tricho )

}
