% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totInfo.R
\name{totInfo}
\alias{totInfo}
\title{totInfo}
\usage{
totInfo(x, taxalist = FALSE)
}
\arguments{
\item{x}{results of the function \link{aggregatoR}}

\item{taxalist}{if \code{TRUE} returns the list of taxa for each taxonomic level.}
}
\description{
\Sexpr[results=rd, stage=render]{ lifecycle::badge("deprecated") }

Function to calculate the overall richness at different taxonomic level.
}
\seealso{
\code{\link{asBiomonitor}}
}
\keyword{totInfo}
