% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleTraits.R, R/traitScaling.R
\name{sampleTraits}
\alias{sampleTraits}
\alias{traitScaling}
\title{traitScaling}
\usage{
sampleTraits(x)

traitScaling(x, traitDB = NULL, dfref = NULL,
  filter_by_distance = NULL)
}
\arguments{
\item{x}{results of function aggregatoR}

\item{traitDB}{a trait data base with a column \code{Taxa} and the other columns
containing the traits. If a trait has several modalities they should be
named as follow: TRAIT_MODALITY.

By default, the data base used is the one from Tachet \emph{et al} (2010) that
can be retrieved from
\href{https://www.freshwaterecology.info/}{freshwaterecology.info} website
(Schmidt-Kloiber & Hering, 2015).}

\item{dfref}{reference database as used in the function aggregatoR.}

\item{filter_by_distance}{filter the results according to the taxonomic distance. Possible values are "pos" , "neg" or a positive integer. See details.}
}
\description{
A function for scaling traits across taxonomic levels.
}
\details{
This function allows to obtain missing traits for a target taxon by taking traits from lower or to upper taxomic levels.
For instance, consider the case where the genus Acroloxus is present in the user dataset and the species Acroloxus lacustris
in the traits database. A simple merge would exclude Acroloxus from the tha analysis since Acroloxus and A. lacustris
would not match. The function traitScaling allows to assign Acroloxus lacustris traits to Acroloxus.
This function works also in the opposite direction. Consider the case where there are no traits for the target taxon and
the target taxon has not been identified at species level. The function traitScaling will assign the traits of the nearest taxonomic level
to the target taxa (e.g. Tanypodinae traits assign to Ablabesmyia monilis). Consider also these examples to understand the behaviour of this
function. For instance Anabolia lombarda is present in the user taxomic dataset while only Anabolia nervosa and Anabolia are present
in the trait database. In this case traitScaling will assing to A. lombarda only the traits of Anabolia.
Moreover, let assume that Coelostoma is present in the user dataset while only Berosus and Crenitis punctatostriata are present in the traits database.
Here traitScaling will assign to Coelostoma the scores of Berosus and C. punctatostriata since they belong to the same family and there are no information at family level.\cr
In case of multiple matches for the target taxa, traitScaling will not average the scores but it will provide all the matches. It is
then straightforward to average the results by using the \link[stats]{aggregate}. \cr
The function traitScaling will measure the taxonomic distance between the target taxa and the taxa used to assign the trait score. This distance
can be negative (e.g. Species to Genus) and positive (e.g. Genus to Species). The distance is measured assigning values as follows:
1 (Species to Genus) , 2 (Species to family), -1 (Family to Genus), etc. traitScaling considers only the
taxonomic levels from Subspecies to Family (Subspecies, Species, Genus, Tribus, Subfamily, Family).
}
\section{Functions}{
\itemize{
\item \code{sampleTraits}: sampleTraits
}}

\note{
USE WITH CAUTION, STILL IN DEVELOPMENT.
}
\examples{
data(macro_ex)

data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
data.ts <- traitScaling( data.agR )

# averaging

data.ts.av <- aggregate( . ~ Taxa , data = data.ts[ , -c(2:5)] , FUN = mean  )

# traits random sampling
data.ts.st <- sampleTraits( data.ts )
}
\seealso{
\link{aggregatoR}
}
