% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tree.R
\docType{data}
\name{Tree}
\alias{Tree}
\title{Taxonomic tree example}
\format{
A data.frame with 5 rows and 6 variables:
}
\usage{
Tree
}
\description{
An example of taxonomic tree to build a reference database suitable for biomonitoR.
}
\keyword{datasets}
