% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whpt.R
\name{whpt}
\alias{whpt}
\title{whpt}
\usage{
whpt(x, taxLev = "Family", type = "ab", method = "aspt", composite = F,
  abucl = c(0, 9, 99, 999))
}
\arguments{
\item{x}{results of aggregatoR function}

\item{taxLev}{currently only the option "Family" is enabled}

\item{type}{presence only ("po") or abundance ("ab")}

\item{method}{possible choices are "aspt", "ntaxa", "bmwp"}

\item{composite}{if T composite families as listed in the details section are used}

\item{abucl}{abundance threshold. Default 0, 9, 99, 999.}
}
\description{
This function calculates WHPT index according to most recent version used in UK.
}
\details{
WHPT is a revision of BMWP and it takes into account the abundances of organisms. The following aggregation is used if composit is set equal to T:

\enumerate{
\item Psychomyiidae (inc. Ecnomidae)
\item Rhyacophilidae (inc. Glossomatidae)
\item Ancylidae (inc. Acroloxidae)
\item Gammaridae (inc. Crangonyctidae)
\item Planariidae (inc. Dugesidae)
\item Hydrobiidae (inc. Bithyniidae)
}

Scores used for whpt calculation can be explored with the function code{\link{showscores}}.
}
\section{Acknowledgements}{
 We thank Carol Fitzpatrick, Richard Chadd, Judy England and Rachel Stubbington for providing us with the most updated WHPT scores and algorithms.
}

\examples{
data(macro_ex)
data.bio <- asBiomonitor(macro_ex)
data.agR <- aggregatoR(data.bio)
data.whpt <- whpt(data.agR, taxLev = "Family", composite = F)
}
\seealso{
\code{\link{asBiomonitor}}
}
\keyword{whpt}
